refactor(market-data): Phase5d extract cold Ensure* module (I-24d)
Extract I-24d cold-path Ensure* handlers and cache-scoped RPC refresh
from market_data bin into src/market_data/cold/.

- ColdHost trait (bin: impl for MarketDataContext)
- ensure_{pump,pumpfun,orca,meteora,raydium}.rs + rpc_refresh.rs
- defer.rs, pump_layout.rs, host.rs
- Bin LOC -3558; behavior unchanged (ControlRequest/Reply, JetStream SSOT)
- phase5d_cold_ensure_in_dedicated_module regression test
