+++ src/bin/market_data.rs
@@ -39,17 +39,16 @@ use uuid::Uuid;
 use ironcrab::config::{Config, MarketDataGeyserCfg, WalletTrackerCfg};
 use ironcrab::ipc::{
     BinData, ConfigUpdate, ConfigUpdateResponse, ConfigUpdateStatus, ControlRequest,
-    ControlRequestKind, ControlResponse, ControlResponseStatus, DexPoolReadiness, ExecutionResult,
-    ExecutionStatus, IntentTier, MarketEvent, MarketEventKind, PoolCacheUpdate,
-    PriorityFeePercentiles, NATIVE_SOL_MINT, POOL_CACHE_UPDATE_METEORA_CPMM_ONCHAIN_MINTS_KEY,
-    POOL_CACHE_UPDATE_METEORA_CPMM_VAULTS_KEY, POOL_CACHE_UPDATE_METEORA_DLMM_ACTIVE_ID_KEY,
-    POOL_CACHE_UPDATE_METEORA_DLMM_BIN_STEP_KEY, POOL_CACHE_UPDATE_METEORA_DLMM_ONCHAIN_MINTS_KEY,
-    POOL_CACHE_UPDATE_METEORA_DLMM_VAULTS_KEY, POOL_CACHE_UPDATE_ORCA_FEE_RATE_KEY,
-    POOL_CACHE_UPDATE_ORCA_LIQUIDITY_KEY, POOL_CACHE_UPDATE_ORCA_ONCHAIN_MINTS_KEY,
-    POOL_CACHE_UPDATE_ORCA_PROTOCOL_FEE_RATE_KEY, POOL_CACHE_UPDATE_ORCA_SQRT_PRICE_KEY,
-    POOL_CACHE_UPDATE_ORCA_TICK_CURRENT_INDEX_KEY, POOL_CACHE_UPDATE_ORCA_TICK_SPACING_KEY,
-    POOL_CACHE_UPDATE_ORCA_TOKEN_A_PROGRAM_KEY, POOL_CACHE_UPDATE_ORCA_TOKEN_B_PROGRAM_KEY,
-    POOL_CACHE_UPDATE_ORCA_WHIRLPOOL_VAULTS_KEY, POOL_CACHE_UPDATE_RAYDIUM_CPMM_VAULTS_KEY,
+    ControlRequestKind, DexPoolReadiness, ExecutionResult, ExecutionStatus, IntentTier,
+    MarketEvent, MarketEventKind, PoolCacheUpdate, PriorityFeePercentiles, NATIVE_SOL_MINT,
+    POOL_CACHE_UPDATE_RAYDIUM_CPMM_VAULTS_KEY,
+};
+use ironcrab::market_data::cold::{
+    cold_path_rpc_refresh_meteora_cpmm_pool_row, cold_path_rpc_refresh_meteora_dlmm_pool_row,
+    cold_path_rpc_refresh_orca_whirlpool_pool_row, handle_ensure_meteora_cpmm_pool_state,
+    handle_ensure_meteora_dlmm_pool_state, handle_ensure_orca_whirlpool_pool_state,
+    handle_ensure_pump_amm_pool_accounts, handle_ensure_pumpfun_bonding_curve,
+    handle_ensure_raydium_amm_pool_state, handle_ensure_raydium_cpmm_pool_state, ColdHost,
 };
 use ironcrab::market_data::publish::{
     account_path_enqueue_core_market_event as publish_enqueue_core_market_event,
@@ -62,6 +61,7 @@ use ironcrab::market_data::publish::{
     AccountPublishSender, PublishHost, MARKET_DATA_ACCOUNT_PUBLISH_QUEUE_CAP,
     MARKET_DATA_ACCOUNT_PUBLISH_WORKER_DISPATCH_QUEUE_CAP,
 };
+
 use ironcrab::market_data::sidefx::{
     md_sidefx_coalesce_burst as sidefx_coalesce_burst,
     md_sidefx_flush_pending_md_state_jobs as sidefx_flush_pending,
@@ -83,9 +83,8 @@ use ironcrab::metrics::{
     geyser_metrics_set_subscription_accounts, geyser_metrics_set_tracked_pinned_accounts,
     inc_market_data_account_high_priority_queue_depth,
     inc_market_data_account_low_priority_queue_depth, inc_market_data_account_worker_queue_depth,
-    inc_market_data_balance_updated_from_cache_total,
-    inc_market_data_discovery_deferred_md_state_pressure_total,
-    inc_market_data_geyser_sync_partial_total, inc_market_data_geyser_sync_skipped_no_delta_total,
+    inc_market_data_balance_updated_from_cache_total, inc_market_data_geyser_sync_partial_total,
+    inc_market_data_geyser_sync_skipped_no_delta_total,
     inc_market_data_geyser_sync_skipped_rate_limit_total,
     inc_market_data_geyser_tracking_enqueue_dropped_total,
     inc_market_data_geyser_tracking_jobs_processed_total,
@@ -130,18 +129,14 @@ use ironcrab::nats::{
     ArbTrackRequestsUpdate, MomentumActivePoolEntry, MomentumActivePoolsUpdate,
     MomentumRemovedPoolEntry, NatsClient, NatsConfig, CONFIG_STREAM_NAME,
     EXECUTION_RESULTS_STREAM_NAME, TOPIC_ARB_TRACK_REQUESTS, TOPIC_CONTROL_REQUESTS,
-    TOPIC_CONTROL_RESPONSES, TOPIC_EXECUTION_RESULTS, TOPIC_MARKET_EVENTS,
-    TOPIC_MOMENTUM_ACTIVE_POOLS, TOPIC_PRIORITY_FEE_SAMPLES, WALLET_SNAPSHOT_STREAM_NAME,
+    TOPIC_EXECUTION_RESULTS, TOPIC_MARKET_EVENTS, TOPIC_MOMENTUM_ACTIVE_POOLS,
+    TOPIC_PRIORITY_FEE_SAMPLES, WALLET_SNAPSHOT_STREAM_NAME,
 };
 use ironcrab::solana::dex::meteora_bin_array_layout::BinArray;
 use ironcrab::solana::dex::meteora_dlmm::METEORA_DLMM_PROGRAM;
 use ironcrab::solana::dex::meteora_swap_builder::MeteoraDlmmSwapBuilder;
-use ironcrab::solana::dex::pumpfun::{BondingCurveState, PumpFunDex};
-use ironcrab::solana::dex::pumpfun_amm::{
-    pump_amm_inferred_sell_ix_account_count, pump_amm_sell_extended_layout_ready,
-    PumpAmmPoolAccountsDiagnostic, PumpAmmSellExtendedReadinessParams, PumpFunAmmDex,
-};
-use ironcrab::solana::dex::raydium::Raydium;
+use ironcrab::solana::dex::pumpfun::PumpFunDex;
+use ironcrab::solana::dex::pumpfun_amm::PumpFunAmmDex;
 use ironcrab::solana::dex_parser::{
     parse_account_update, parse_transaction_update_with_pool_lookup, DexType, OrcaPoolInfo,
     ParsedDexEvent,
@@ -196,8 +191,6 @@ const MARKET_DATA_MD_STATE_STALL_WINDOW: Duration = Duration::from_secs(60);
 const MARKET_DATA_MD_STATE_STALL_EXIT_AFTER: Duration = Duration::from_secs(120);
 /// PR234: queue depth fraction treated as saturation for md-state liveness.
 const MARKET_DATA_MD_STATE_STALL_QUEUE_FRAC: f64 = 0.95;
-/// PR235: defer cold-path discovery when md-state queue exceeds this fraction of cap.
-const MARKET_DATA_MD_STATE_DISCOVERY_DEFER_QUEUE_FRAC: f64 = 0.75;
 /// PR235: when false, md-state liveness records stalls but does not exit(1) for systemd restart.
 const MARKET_DATA_MD_STATE_STALL_EXIT_ENABLED: bool = false;
 
@@ -207,12 +200,10 @@ const MARKET_DATA_MD_STATE_STALL_EXIT_ENABLED: bool = false;
 const EXECUTION_RESULT_DEDUP_CAPACITY: usize = 4096;
 
 // LivePoolCache - MASTER Cache (Single Source of Truth)
+#[allow(unused_imports)]
 use ironcrab::execution::live_pool_cache::{
-    meteora_cpmm_readiness_for_pool_cache_update, meteora_dlmm_readiness_for_pool_cache_update,
-    orca_readiness_for_pool_cache_update, parse_pool_account,
-    raydium_amm_readiness_for_pool_cache_update, CachedPoolState, LivePoolCache, MeteoraCpmmState,
-    MeteoraState, OrcaWhirlpoolState, PumpAmmState, PumpFunState, RaydiumAmmState,
-    RaydiumCpmmState,
+    parse_pool_account, CachedPoolState, LivePoolCache, MeteoraCpmmState, MeteoraState,
+    OrcaWhirlpoolState, PumpAmmState, PumpFunState, RaydiumAmmState, RaydiumCpmmState,
 };
 
 // P1 Crash Isolation: Systemd Watchdog support
@@ -842,106 +833,6 @@ fn raydium_cpmm_vaults_for_pool_cache_update(s: &RaydiumCpmmState) -> String {
     format!("{first_vault},{second_vault}")
 }
 
-/// Meteora CPMM: same normalized base/quote vault ordering as Raydium CPMM for JetStream metadata.
-fn meteora_cpmm_vaults_for_pool_cache_update(s: &MeteoraCpmmState) -> String {
-    let sol = Pubkey::from_str(NATIVE_SOL_MINT).unwrap_or_default();
-    let (first_vault, second_vault) = if s.token_1_mint == sol {
-        (s.token_0_vault, s.token_1_vault)
-    } else if s.token_0_mint == sol {
-        (s.token_1_vault, s.token_0_vault)
-    } else {
-        (s.token_0_vault, s.token_1_vault)
-    };
-    format!("{first_vault},{second_vault}")
-}
-
-/// On-chain `token_0_mint,token_1_mint` for SLAVE bootstrap when JetStream uses normalized base/quote.
-fn meteora_cpmm_onchain_mints_for_pool_cache_update(s: &MeteoraCpmmState) -> String {
-    format!("{},{}", s.token_0_mint, s.token_1_mint)
-}
-
-/// On-chain `token_mint_a,token_mint_b` for SLAVE bootstrap when JetStream uses normalized base/quote.
-fn orca_onchain_mints_for_pool_cache_update(s: &OrcaWhirlpoolState) -> String {
-    format!("{},{}", s.token_mint_a, s.token_mint_b)
-}
-
-/// Orca Whirlpool: PoolCacheUpdate metadata keys for SLAVE bootstrap (BalanceUpdated + PoolDiscovered).
-fn orca_metadata_for_pool_cache_update(s: &OrcaWhirlpoolState) -> HashMap<String, String> {
-    let mut meta = HashMap::new();
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_WHIRLPOOL_VAULTS_KEY.to_string(),
-        format!("{},{}", s.token_vault_a, s.token_vault_b),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_ONCHAIN_MINTS_KEY.to_string(),
-        orca_onchain_mints_for_pool_cache_update(s),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_TICK_CURRENT_INDEX_KEY.to_string(),
-        s.tick_current_index.to_string(),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_TICK_SPACING_KEY.to_string(),
-        s.tick_spacing.to_string(),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_SQRT_PRICE_KEY.to_string(),
-        s.sqrt_price.to_string(),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_LIQUIDITY_KEY.to_string(),
-        s.liquidity.to_string(),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_FEE_RATE_KEY.to_string(),
-        s.fee_rate.to_string(),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_ORCA_PROTOCOL_FEE_RATE_KEY.to_string(),
-        s.protocol_fee_rate.to_string(),
-    );
-    if let Some(p) = s.token_a_program {
-        meta.insert(
-            POOL_CACHE_UPDATE_ORCA_TOKEN_A_PROGRAM_KEY.to_string(),
-            p.to_string(),
-        );
-    }
-    if let Some(p) = s.token_b_program {
-        meta.insert(
-            POOL_CACHE_UPDATE_ORCA_TOKEN_B_PROGRAM_KEY.to_string(),
-            p.to_string(),
-        );
-    }
-    meta
-}
-
-/// On-chain `token_x_mint,token_y_mint` for SLAVE bootstrap when JetStream uses normalized base/quote.
-fn meteora_dlmm_onchain_mints_for_pool_cache_update(s: &MeteoraState) -> String {
-    format!("{},{}", s.token_x_mint, s.token_y_mint)
-}
-
-/// Meteora DLMM: vault pubkeys + static pool fields for JetStream / SLAVE bootstrap.
-fn meteora_dlmm_metadata_for_pool_cache_update(s: &MeteoraState) -> HashMap<String, String> {
-    let mut meta = HashMap::new();
-    meta.insert(
-        POOL_CACHE_UPDATE_METEORA_DLMM_VAULTS_KEY.to_string(),
-        format!("{},{}", s.reserve_x, s.reserve_y),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_METEORA_DLMM_ONCHAIN_MINTS_KEY.to_string(),
-        meteora_dlmm_onchain_mints_for_pool_cache_update(s),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_METEORA_DLMM_ACTIVE_ID_KEY.to_string(),
-        s.active_id.to_string(),
-    );
-    meta.insert(
-        POOL_CACHE_UPDATE_METEORA_DLMM_BIN_STEP_KEY.to_string(),
-        s.bin_step.to_string(),
-    );
-    meta
-}
-
 /// JetStream readiness for Raydium CPMM (SOL-aware base/quote): single source for BalanceUpdated and PoolDiscovered.
 fn raydium_cpmm_readiness_for_pool_cache_update(s: &RaydiumCpmmState) -> DexPoolReadiness {
     let sol = Pubkey::from_str(NATIVE_SOL_MINT).unwrap_or_default();
@@ -961,194 +852,6 @@ fn raydium_cpmm_readiness_for_pool_cache_update(s: &RaydiumCpmmState) -> DexPool
     } else {
         DexPoolReadiness::Observed
     }
-}
-
-/// PumpSwap SELL-layout contract for JetStream / SLAVE SSOT.
-///
-/// - Base-only SELL stays ready as long as the underlying refresh/observation says the base layout is usable.
-/// - Extended SELL is ready when pool `third_meta` (+ fee tails when required) is known; #21/#22 are derived at build.
-#[allow(clippy::too_many_arguments)]
-fn pump_amm_sell_layout_publish_state(
-    sell_requires_extended: bool,
-    sell_cashback_third_meta: Option<Pubkey>,
-    sell_extended_tail_0: Option<Pubkey>,
-    sell_extended_tail_1: Option<Pubkey>,
-    sell_extended_fee_tail_0: Option<Pubkey>,
-    sell_extended_fee_tail_1: Option<Pubkey>,
-    sell_requires_fee_tail: bool,
-    sell_requires_pre_fee_metas: bool,
-    sell_pre_fee_meta_1: Option<Pubkey>,
-    base_layout_ready: bool,
-) -> (bool, DexPoolReadiness) {
-    let sell_layout_ready = if sell_requires_extended {
-        let _ = (sell_extended_tail_0, sell_extended_tail_1);
-        pump_amm_sell_extended_layout_ready(PumpAmmSellExtendedReadinessParams {
-            sell_requires_extended: true,
-            third_meta: sell_cashback_third_meta,
-            fee_tail_0: sell_extended_fee_tail_0,
-            fee_tail_1: sell_extended_fee_tail_1,
-            sell_requires_fee_tail,
-            sell_requires_pre_fee_metas,
-            sell_pre_fee_meta_1,
-        }) && base_layout_ready
-    } else {
-        base_layout_ready
-    };
-    let dex_readiness = if sell_layout_ready {
-        DexPoolReadiness::Ready
-    } else {
-        DexPoolReadiness::Partial
-    };
-    (sell_layout_ready, dex_readiness)
-}
-
-/// Resolve PumpSwap SELL-layout state for the EnsurePumpAmmPoolAccounts publish path.
-///
-/// `force_refresh=true` is authoritative and must be able to override stale monotonic cache hints.
-#[allow(clippy::too_many_arguments, clippy::type_complexity)]
-fn pump_amm_sell_layout_state_for_ensure_publish(
-    force_refresh: bool,
-    cached_requires_extended: bool,
-    cached_third_meta: Option<Pubkey>,
-    cached_tail_0: Option<Pubkey>,
-    cached_tail_1: Option<Pubkey>,
-    cached_fee_tail_0: Option<Pubkey>,
-    cached_fee_tail_1: Option<Pubkey>,
-    cached_requires_fee_tail: bool,
-    cached_requires_pre_fee_metas: bool,
-    cached_pre_fee_meta_1: Option<Pubkey>,
-    refresh_requires_extended: bool,
-    refresh_third_meta: Option<Pubkey>,
-    refresh_tail_0: Option<Pubkey>,
-    refresh_tail_1: Option<Pubkey>,
-    refresh_fee_tail_0: Option<Pubkey>,
-    refresh_fee_tail_1: Option<Pubkey>,
-    refresh_requires_fee_tail: bool,
-    refresh_requires_pre_fee_metas: bool,
-    refresh_pre_fee_meta_1: Option<Pubkey>,
-    refresh_layout_ready: bool,
-) -> (
-    bool,
-    Option<Pubkey>,
-    Option<Pubkey>,
-    Option<Pubkey>,
-    Option<Pubkey>,
-    Option<Pubkey>,
-    bool,
-    bool,
-    Option<Pubkey>,
-    bool,
-    DexPoolReadiness,
-) {
-    let (
-        effective_requires_extended,
-        effective_third_meta,
-        effective_tail_0,
-        effective_tail_1,
-        effective_fee_tail_0,
-        effective_fee_tail_1,
-        effective_requires_fee_tail,
-        effective_requires_pre_fee_metas,
-        effective_pre_fee_meta_1,
-        base_layout_ready,
-    ) = if force_refresh {
-        (
-            refresh_requires_extended || cached_requires_extended,
-            refresh_third_meta.filter(|p| *p != Pubkey::default()),
-            refresh_tail_0.filter(|p| *p != Pubkey::default()),
-            refresh_tail_1.filter(|p| *p != Pubkey::default()),
-            refresh_fee_tail_0
-                .or(cached_fee_tail_0)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_fee_tail_1
-                .or(cached_fee_tail_1)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_requires_fee_tail || cached_requires_fee_tail,
-            refresh_requires_pre_fee_metas,
-            if refresh_requires_pre_fee_metas {
-                refresh_pre_fee_meta_1
-                    .or(cached_pre_fee_meta_1)
-                    .filter(|p| *p != Pubkey::default())
-            } else {
-                None
-            },
-            refresh_layout_ready,
-        )
-    } else {
-        (
-            cached_requires_extended || refresh_requires_extended,
-            refresh_third_meta
-                .or(cached_third_meta)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_tail_0
-                .or(cached_tail_0)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_tail_1
-                .or(cached_tail_1)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_fee_tail_0
-                .or(cached_fee_tail_0)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_fee_tail_1
-                .or(cached_fee_tail_1)
-                .filter(|p| *p != Pubkey::default()),
-            cached_requires_fee_tail || refresh_requires_fee_tail,
-            cached_requires_pre_fee_metas || refresh_requires_pre_fee_metas,
-            refresh_pre_fee_meta_1
-                .or(cached_pre_fee_meta_1)
-                .filter(|p| *p != Pubkey::default()),
-            refresh_layout_ready,
-        )
-    };
-    let (sell_layout_ready, dex_readiness) = pump_amm_sell_layout_publish_state(
-        effective_requires_extended,
-        effective_third_meta,
-        effective_tail_0,
-        effective_tail_1,
-        effective_fee_tail_0,
-        effective_fee_tail_1,
-        effective_requires_fee_tail,
-        effective_requires_pre_fee_metas,
-        effective_pre_fee_meta_1,
-        base_layout_ready,
-    );
-    (
-        effective_requires_extended,
-        effective_third_meta,
-        effective_tail_0,
-        effective_tail_1,
-        effective_fee_tail_0,
-        effective_fee_tail_1,
-        effective_requires_fee_tail,
-        effective_requires_pre_fee_metas,
-        effective_pre_fee_meta_1,
-        sell_layout_ready,
-        dex_readiness,
-    )
-}
-
-/// Control-response contract for PumpSwap EnsurePumpAmmPoolAccounts.
-///
-/// Only the authoritative `force_refresh=true` path may turn a successful JetStream publish into an
-/// error when the resolved SELL layout is still not ready.
-fn pump_amm_control_response_for_ensure_publish(
-    force_refresh: bool,
-    jetstream_ok: bool,
-    sell_layout_ready: bool,
-) -> (ControlResponseStatus, Option<String>) {
-    if !jetstream_ok {
-        return (
-            ControlResponseStatus::Error,
-            Some("JetStream publish failed".to_string()),
-        );
-    }
-    if force_refresh && !sell_layout_ready {
-        return (
-            ControlResponseStatus::Error,
-            Some("authoritative PumpSwap SELL layout unresolved after force_refresh".to_string()),
-        );
-    }
-    (ControlResponseStatus::Ok, None)
 }
 
 /// PR-B: explicit Geyser account subscription / LRU (also loaded from `[market_data_geyser]` in config.toml).
@@ -2419,6 +2122,28 @@ impl TrackWorkerContext for MarketDataContext {
             rows.push((*pk, ConsumerId::Wallet, None));
         }
         rows
+    }
+}
+
+impl ColdHost for MarketDataContext {
+    fn run_id(&self) -> &str {
+        &self.run_id
+    }
+
+    fn nats(&self) -> Option<&NatsClient> {
+        self.nats.as_ref()
+    }
+
+    fn live_pool_cache(&self) -> &LivePoolCache {
+        self.live_pool_cache.as_ref()
+    }
+
+    fn live_pool_cache_arc(&self) -> Arc<LivePoolCache> {
+        Arc::clone(&self.live_pool_cache)
+    }
+
+    fn raydium_serum_fetched_insert(&self, pool_addr: Pubkey) {
+        self.raydium_serum_fetched.write().insert(pool_addr);
     }
 }
 
@@ -4849,194 +4574,6 @@ async fn handle_wallet_bootstrap_raydium_cpmm_verify_for_mint(
     }
 }
 
-/// Result of one cache-scoped Meteora CPMM RPC refresh (pool + vault balances).
-struct MeteoraCpmmRpcRefreshRowResult {
-    pool_addr: Pubkey,
-    readiness: DexPoolReadiness,
-    /// `true` when a non-`Observed` [`DexPoolReadiness`] update was published to JetStream.
-    jetstream_published: bool,
-}
-
-/// Cold-path only: refresh one Meteora CPMM row already present in MASTER cache — no global scan.
-async fn cold_path_rpc_refresh_meteora_cpmm_pool_row(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint: &Pubkey,
-    pool_addr: Pubkey,
-    mut state: MeteoraCpmmState,
-) -> Option<MeteoraCpmmRpcRefreshRowResult> {
-    let meteora_cpmm_program = Pubkey::from_str(METEORA_CPMM).expect("METEORA_CPMM constant");
-    let sol = Pubkey::from_str(NATIVE_SOL_MINT).unwrap_or_default();
-
-    let pool_acc = match rpc.get_account_opt_retry(&pool_addr).await {
-        Ok(Some(acc)) => acc,
-        Ok(None) | Err(_) => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Meteora CPMM RPC refresh: pool account fetch miss, skip"
-            );
-            return None;
-        }
-    };
-
-    let parsed_state = match parse_pool_account(&meteora_cpmm_program, &pool_acc.data) {
-        Some(CachedPoolState::MeteoraCpmm(s)) => s,
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                data_len = pool_acc.data.len(),
-                "Meteora CPMM RPC refresh: parse pool failed, skip"
-            );
-            return None;
-        }
-    };
-
-    state.token_0_mint = parsed_state.token_0_mint;
-    state.token_1_mint = parsed_state.token_1_mint;
-    state.token_0_vault = parsed_state.token_0_vault;
-    state.token_1_vault = parsed_state.token_1_vault;
-    state.amm_config = parsed_state.amm_config;
-    state.observation_key = parsed_state.observation_key;
-    state.token_0_program = parsed_state.token_0_program;
-    state.token_1_program = parsed_state.token_1_program;
-    state.mint_0_decimals = parsed_state.mint_0_decimals;
-    state.mint_1_decimals = parsed_state.mint_1_decimals;
-    state.status = parsed_state.status;
-
-    if state.token_0_mint != *base_mint && state.token_1_mint != *base_mint {
-        return None;
-    }
-
-    let vault0 = state.token_0_vault;
-    let vault1 = state.token_1_vault;
-    let bal0 = match rpc.get_account_opt_retry(&vault0).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let bal1 = match rpc.get_account_opt_retry(&vault1).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let (b0, b1) = match (bal0, bal1) {
-        (Some(a), Some(b)) => (a, b),
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Meteora CPMM RPC refresh: vault balance RPC incomplete, skip"
-            );
-            return None;
-        }
-    };
-
-    state.reserve_0 = b0;
-    state.reserve_1 = b1;
-
-    ctx.live_pool_cache
-        .upsert(pool_addr, CachedPoolState::MeteoraCpmm(state.clone()), 0);
-
-    let readiness = meteora_cpmm_readiness_for_pool_cache_update(&state);
-    ctx.live_pool_cache
-        .merge_meteora_cpmm_pool_readiness(pool_addr, readiness);
-
-    if readiness == DexPoolReadiness::Observed {
-        debug!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            mint = %base_mint,
-            "Meteora CPMM RPC refresh: reserves still degenerate (Observed), no JetStream publish"
-        );
-        return Some(MeteoraCpmmRpcRefreshRowResult {
-            pool_addr,
-            readiness,
-            jetstream_published: false,
-        });
-    }
-
-    let (pub_base_mint, pub_quote_mint, base_r, quote_r) = if state.token_1_mint == sol {
-        (state.token_0_mint, state.token_1_mint, b0, b1)
-    } else if state.token_0_mint == sol {
-        (state.token_1_mint, state.token_0_mint, b1, b0)
-    } else {
-        (state.token_0_mint, state.token_1_mint, b0, b1)
-    };
-
-    let jetstream_ok = if let Some(ref nats) = ctx.nats {
-        let mut balance_update = PoolCacheUpdate::new_balance_updated(
-            "market-data",
-            BUILD_VERSION,
-            run_id,
-            pool_addr.to_string(),
-            "meteora_cpmm".to_string(),
-            pub_base_mint.to_string(),
-            pub_quote_mint.to_string(),
-            base_r,
-            quote_r,
-            0,
-        );
-        let mut meta = std::collections::HashMap::new();
-        meta.insert(
-            POOL_CACHE_UPDATE_METEORA_CPMM_VAULTS_KEY.to_string(),
-            meteora_cpmm_vaults_for_pool_cache_update(&state),
-        );
-        meta.insert(
-            POOL_CACHE_UPDATE_METEORA_CPMM_ONCHAIN_MINTS_KEY.to_string(),
-            meteora_cpmm_onchain_mints_for_pool_cache_update(&state),
-        );
-        balance_update.metadata = Some(meta);
-        balance_update.set_dex_readiness_in_metadata(readiness);
-        let subject = pool_subject(&pool_addr.to_string());
-        match nats.jetstream_publish(&subject, &balance_update).await {
-            Ok(true) => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    mint = %base_mint,
-                    readiness = ?readiness,
-                    "Meteora CPMM RPC refresh: published PoolCacheUpdate::BalanceUpdated to JetStream"
-                );
-                true
-            }
-            Ok(false) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    "Meteora CPMM RPC refresh: JetStream publish failed (timeout or drop)"
-                );
-                false
-            }
-            Err(e) => {
-                warn!(
-                    error = %e,
-                    request_id = %request_id,
-                    "Meteora CPMM RPC refresh: Failed to publish PoolCacheUpdate to JetStream"
-                );
-                false
-            }
-        }
-    } else {
-        false
-    };
-
-    if !jetstream_ok {
-        warn!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            "Meteora CPMM RPC refresh: MASTER cache updated but JetStream publish failed — SSOT drift risk"
-        );
-    }
-
-    Some(MeteoraCpmmRpcRefreshRowResult {
-        pool_addr,
-        readiness,
-        jetstream_published: jetstream_ok,
-    })
-}
-
 /// Cold-path only: Meteora CPMM promotion for one wallet mint. Uses **only** pools already present
 /// in [`LivePoolCache`] (from Geyser); no global scan. RPC: per-pool `get_account` + per-vault
 /// balance reads — bounded by the number of matching cache rows.
@@ -5090,187 +4627,11 @@ async fn handle_wallet_bootstrap_meteora_cpmm_verify_for_mint(
             break;
         }
 
-        let _ = cold_path_rpc_refresh_meteora_cpmm_pool_row(
+        cold_path_rpc_refresh_meteora_cpmm_pool_row(
             ctx, rpc, run_id, request_id, base_mint, pool_addr, state,
         )
         .await;
     }
-}
-
-/// Result of one cache-scoped Orca Whirlpool RPC refresh (pool account + vault balances).
-struct OrcaWhirlpoolRpcRefreshRowResult {
-    pool_addr: Pubkey,
-    readiness: DexPoolReadiness,
-    /// `true` when a non-`Observed` [`DexPoolReadiness`] update was published to JetStream.
-    jetstream_published: bool,
-}
-
-/// Cold-path only: refresh one Orca Whirlpool row already present in MASTER cache — same RPC
-/// pattern as wallet bootstrap (no global scan). Updates MASTER, merges readiness, publishes
-/// JetStream when readiness is not `Observed`.
-///
-/// Returns `None` when this pool row is skipped (fetch/parse/mint mismatch/incomplete vaults).
-async fn cold_path_rpc_refresh_orca_whirlpool_pool_row(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint: &Pubkey,
-    pool_addr: Pubkey,
-    mut state: OrcaWhirlpoolState,
-) -> Option<OrcaWhirlpoolRpcRefreshRowResult> {
-    let orca_program = Pubkey::from_str(ORCA_WHIRLPOOL).expect("ORCA_WHIRLPOOL constant");
-
-    let pool_acc = match rpc.get_account_opt_retry(&pool_addr).await {
-        Ok(Some(acc)) => acc,
-        Ok(None) | Err(_) => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Orca Whirlpool RPC refresh: pool account fetch miss, skip"
-            );
-            return None;
-        }
-    };
-
-    let parsed_state = match parse_pool_account(&orca_program, &pool_acc.data) {
-        Some(CachedPoolState::Orca(s)) => s,
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                data_len = pool_acc.data.len(),
-                "Orca Whirlpool RPC refresh: parse pool failed, skip"
-            );
-            return None;
-        }
-    };
-
-    state.token_mint_a = parsed_state.token_mint_a;
-    state.token_mint_b = parsed_state.token_mint_b;
-    state.token_vault_a = parsed_state.token_vault_a;
-    state.token_vault_b = parsed_state.token_vault_b;
-    state.tick_current_index = parsed_state.tick_current_index;
-    state.sqrt_price = parsed_state.sqrt_price;
-    state.liquidity = parsed_state.liquidity;
-    state.fee_rate = parsed_state.fee_rate;
-    state.protocol_fee_rate = parsed_state.protocol_fee_rate;
-    state.tick_spacing = parsed_state.tick_spacing;
-
-    if state.token_mint_a != *base_mint && state.token_mint_b != *base_mint {
-        return None;
-    }
-
-    let vault_a = state.token_vault_a;
-    let vault_b = state.token_vault_b;
-    let bal_a = match rpc.get_account_opt_retry(&vault_a).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let bal_b = match rpc.get_account_opt_retry(&vault_b).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let (ba, bb) = match (bal_a, bal_b) {
-        (Some(a), Some(b)) => (a, b),
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Orca Whirlpool RPC refresh: vault balance RPC incomplete, skip"
-            );
-            return None;
-        }
-    };
-
-    state.vault_a_balance = Some(ba);
-    state.vault_b_balance = Some(bb);
-
-    ctx.live_pool_cache
-        .upsert(pool_addr, CachedPoolState::Orca(state.clone()), 0);
-
-    let readiness = orca_readiness_for_pool_cache_update(&state);
-    ctx.live_pool_cache
-        .merge_orca_pool_readiness(pool_addr, readiness);
-
-    if readiness == DexPoolReadiness::Observed {
-        debug!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            mint = %base_mint,
-            "Orca Whirlpool RPC refresh: still Observed after RPC, no JetStream publish"
-        );
-        return Some(OrcaWhirlpoolRpcRefreshRowResult {
-            pool_addr,
-            readiness,
-            jetstream_published: false,
-        });
-    }
-
-    let (pub_base_mint, pub_quote_mint, base_r, quote_r) =
-        (state.token_mint_a, state.token_mint_b, ba, bb);
-
-    let jetstream_ok = if let Some(ref nats) = ctx.nats {
-        let mut balance_update = PoolCacheUpdate::new_balance_updated(
-            "market-data",
-            BUILD_VERSION,
-            run_id,
-            pool_addr.to_string(),
-            "orca".to_string(),
-            pub_base_mint.to_string(),
-            pub_quote_mint.to_string(),
-            base_r,
-            quote_r,
-            0,
-        );
-        balance_update.metadata = Some(orca_metadata_for_pool_cache_update(&state));
-        balance_update.set_dex_readiness_in_metadata(readiness);
-        let subject = pool_subject(&pool_addr.to_string());
-        match nats.jetstream_publish(&subject, &balance_update).await {
-            Ok(true) => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    mint = %base_mint,
-                    readiness = ?readiness,
-                    "Orca Whirlpool RPC refresh: published PoolCacheUpdate::BalanceUpdated to JetStream"
-                );
-                true
-            }
-            Ok(false) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    "Orca Whirlpool RPC refresh: JetStream publish failed (timeout or drop)"
-                );
-                false
-            }
-            Err(e) => {
-                warn!(
-                    error = %e,
-                    request_id = %request_id,
-                    "Orca Whirlpool RPC refresh: Failed to publish PoolCacheUpdate to JetStream"
-                );
-                false
-            }
-        }
-    } else {
-        false
-    };
-
-    if !jetstream_ok {
-        warn!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            "Orca Whirlpool RPC refresh: MASTER cache updated but JetStream publish failed — SSOT drift risk"
-        );
-    }
-
-    Some(OrcaWhirlpoolRpcRefreshRowResult {
-        pool_addr,
-        readiness,
-        jetstream_published: jetstream_ok,
-    })
 }
 
 /// Cold-path only: Orca Whirlpool promotion for one wallet mint. Uses **only** pools already present
@@ -5330,2030 +4691,13 @@ async fn handle_wallet_bootstrap_orca_whirlpool_verify_for_mint(
             break;
         }
 
-        let _ = cold_path_rpc_refresh_orca_whirlpool_pool_row(
+        cold_path_rpc_refresh_orca_whirlpool_pool_row(
             ctx, rpc, run_id, request_id, base_mint, pool_addr, state,
         )
         .await;
     }
 }
 
-/// I-24d / Scope 20: Cold-path Orca recovery — cache-scoped RPC refresh, JetStream SSOT, ControlResponse.
-async fn handle_ensure_orca_whirlpool_pool_state(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsureOrcaWhirlpoolPoolState: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pool_hint_pk = pool_address_hint.and_then(|s| Pubkey::from_str(s).ok());
-
-    if !force_refresh {
-        if let Some(pool_pk) = pool_hint_pk {
-            if ctx.live_pool_cache.orca_pool_explicitly_ready(&pool_pk) {
-                info!(
-                    request_id = %request_id,
-                    base_mint = %base_mint_str,
-                    pool = %pool_pk,
-                    "EnsureOrcaWhirlpoolPoolState: cache hit explicit Ready (skip RPC)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Ok,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        } else if ctx
-            .live_pool_cache
-            .base_mint_has_explicit_orca_ready_pool(&base_mint)
-        {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureOrcaWhirlpoolPoolState: mint already has explicit Ready Orca pool (skip RPC)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Ok,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool_address_hint = ?pool_address_hint,
-            "EnsureOrcaWhirlpoolPoolState: force_refresh — always running RPC refresh path (no cache-first short-circuit)"
-        );
-    }
-
-    let mut candidate_pools: Vec<(Pubkey, OrcaWhirlpoolState)> = Vec::new();
-    if let Some(pool_pk) = pool_hint_pk {
-        match ctx.live_pool_cache.get(&pool_pk) {
-            Some(CachedPoolState::Orca(s)) => {
-                if s.token_mint_a != base_mint && s.token_mint_b != base_mint {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %pool_pk,
-                        base_mint = %base_mint_str,
-                        "EnsureOrcaWhirlpoolPoolState: pool_address_hint Orca row does not list base_mint"
-                    );
-                    if let Some(ref nats) = ctx.nats {
-                        publish_control_response(
-                            nats,
-                            &ctx.run_id,
-                            request_id,
-                            ControlResponseStatus::Error,
-                            Some(pool_pk.to_string()),
-                            Some("pool hint mint mismatch".to_string()),
-                        )
-                        .await;
-                    }
-                    return;
-                }
-                candidate_pools.push((pool_pk, s));
-            }
-            Some(_) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    "EnsureOrcaWhirlpoolPoolState: pool_address_hint is not an Orca Whirlpool row"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Error,
-                        Some(pool_pk.to_string()),
-                        Some("pool hint is not orca whirlpool in cache".to_string()),
-                    )
-                    .await;
-                }
-                return;
-            }
-            None => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    base_mint = %base_mint_str,
-                    "EnsureOrcaWhirlpoolPoolState: pool hint not in LivePoolCache (NotFound)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::NotFound,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        }
-    } else {
-        candidate_pools = ctx
-            .live_pool_cache
-            .orca_whirlpool_pools_for_mint(&base_mint);
-        if candidate_pools.is_empty() {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureOrcaWhirlpoolPoolState: no Whirlpool rows in LivePoolCache for mint (NotFound)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::NotFound,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    }
-
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        pool_address_hint = ?pool_address_hint,
-        candidates = candidate_pools.len(),
-        force_refresh = %force_refresh,
-        "EnsureOrcaWhirlpoolPoolState: starting cache-scoped Orca RPC refresh"
-    );
-
-    // Terminal `ControlResponse::Ok` only when JetStream carried explicit Orca `Ready` (matches
-    // execution-engine `wait_for_orca_whirlpool_slave_after_recovery` — Bug #36: Partial ≠ ready).
-    let mut terminal_ok_pool: Option<Pubkey> = None;
-    // Best-effort: on-chain Ready but JetStream publish failed — report only if no other candidate succeeds.
-    let mut ready_jetstream_failed_pool: Option<Pubkey> = None;
-
-    for (pool_addr, state) in candidate_pools {
-        if let Some(row) = cold_path_rpc_refresh_orca_whirlpool_pool_row(
-            ctx, rpc, run_id, request_id, &base_mint, pool_addr, state,
-        )
-        .await
-        {
-            match (row.readiness, row.jetstream_published) {
-                (DexPoolReadiness::Ready, true) => {
-                    terminal_ok_pool = Some(row.pool_addr);
-                    break;
-                }
-                (DexPoolReadiness::Ready, false) => {
-                    ready_jetstream_failed_pool = Some(row.pool_addr);
-                }
-                (DexPoolReadiness::Partial, true) => {
-                    debug!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureOrcaWhirlpoolPoolState: Partial published to JetStream — scan remaining candidates for Ready"
-                    );
-                }
-                (DexPoolReadiness::Partial, false) => {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureOrcaWhirlpoolPoolState: Partial row MASTER-updated but JetStream publish failed"
-                    );
-                }
-                (DexPoolReadiness::Observed, _) => {}
-            }
-        }
-    }
-
-    if let Some(pool_pk) = terminal_ok_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                readiness = ?DexPoolReadiness::Ready,
-                jetstream = "published",
-                "EnsureOrcaWhirlpoolPoolState: terminal ok (Orca Ready + JetStream publish)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Ok,
-                Some(pool_str),
-                None,
-            )
-            .await;
-        }
-        return;
-    }
-
-    if let Some(pool_pk) = ready_jetstream_failed_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            let msg = "JetStream publish failed".to_string();
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                error = %msg,
-                "EnsureOrcaWhirlpoolPoolState: terminal error (Orca Ready on MASTER but JetStream publish failed)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Error,
-                Some(pool_str),
-                Some(msg),
-            )
-            .await;
-        }
-        return;
-    }
-
-    warn!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        "EnsureOrcaWhirlpoolPoolState: RPC refresh did not yield Orca Ready + JetStream publish (Error)"
-    );
-    if let Some(ref nats) = ctx.nats {
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            ControlResponseStatus::Error,
-            None,
-            Some("orca rpc refresh did not produce jetstream-ready state".to_string()),
-        )
-        .await;
-    }
-}
-
-/// Result of one cache-scoped Meteora DLMM RPC refresh (LB pair + vault balances).
-struct MeteoraDlmmRpcRefreshRowResult {
-    pool_addr: Pubkey,
-    readiness: DexPoolReadiness,
-    /// `true` when a non-`Observed` [`DexPoolReadiness`] update was published to JetStream.
-    jetstream_published: bool,
-}
-
-/// Cold-path only: refresh one Meteora DLMM row already present in MASTER cache — same RPC
-/// pattern as wallet bootstrap (no global scan). Updates MASTER, merges readiness, publishes
-/// JetStream when readiness is not `Observed`.
-///
-/// Returns `None` when this pool row is skipped (fetch/parse/mint mismatch/incomplete vaults).
-async fn cold_path_rpc_refresh_meteora_dlmm_pool_row(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint: &Pubkey,
-    pool_addr: Pubkey,
-    mut state: MeteoraState,
-) -> Option<MeteoraDlmmRpcRefreshRowResult> {
-    let dlmm_program = Pubkey::from_str(METEORA_DLMM).expect("METEORA_DLMM constant");
-
-    let pool_acc = match rpc.get_account_opt_retry(&pool_addr).await {
-        Ok(Some(acc)) => acc,
-        Ok(None) | Err(_) => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Meteora DLMM RPC refresh: pool account fetch miss, skip"
-            );
-            return None;
-        }
-    };
-
-    let parsed_state = match parse_pool_account(&dlmm_program, &pool_acc.data) {
-        Some(CachedPoolState::Meteora(s)) => s,
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                data_len = pool_acc.data.len(),
-                "Meteora DLMM RPC refresh: parse pool failed, skip"
-            );
-            return None;
-        }
-    };
-
-    state.token_x_mint = parsed_state.token_x_mint;
-    state.token_y_mint = parsed_state.token_y_mint;
-    state.reserve_x = parsed_state.reserve_x;
-    state.reserve_y = parsed_state.reserve_y;
-    state.active_id = parsed_state.active_id;
-    state.bin_step = parsed_state.bin_step;
-
-    if state.token_x_mint != *base_mint && state.token_y_mint != *base_mint {
-        return None;
-    }
-
-    let vx = state.reserve_x;
-    let vy = state.reserve_y;
-    let bal_x = match rpc.get_account_opt_retry(&vx).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let bal_y = match rpc.get_account_opt_retry(&vy).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let (bx, by) = match (bal_x, bal_y) {
-        (Some(a), Some(b)) => (a, b),
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Meteora DLMM RPC refresh: vault balance RPC incomplete, skip"
-            );
-            return None;
-        }
-    };
-
-    state.reserve_x_balance = Some(bx);
-    state.reserve_y_balance = Some(by);
-
-    ctx.live_pool_cache
-        .upsert(pool_addr, CachedPoolState::Meteora(state.clone()), 0);
-
-    let readiness = meteora_dlmm_readiness_for_pool_cache_update(&state);
-    ctx.live_pool_cache
-        .merge_meteora_dlmm_pool_readiness(pool_addr, readiness);
-
-    if readiness == DexPoolReadiness::Observed {
-        debug!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            mint = %base_mint,
-            "Meteora DLMM RPC refresh: still Observed after RPC, no JetStream publish"
-        );
-        return Some(MeteoraDlmmRpcRefreshRowResult {
-            pool_addr,
-            readiness,
-            jetstream_published: false,
-        });
-    }
-
-    let (pub_base_mint, pub_quote_mint, base_r, quote_r) =
-        (state.token_x_mint, state.token_y_mint, bx, by);
-
-    let jetstream_ok = if let Some(ref nats) = ctx.nats {
-        let mut balance_update = PoolCacheUpdate::new_balance_updated(
-            "market-data",
-            BUILD_VERSION,
-            run_id,
-            pool_addr.to_string(),
-            "meteora_dlmm".to_string(),
-            pub_base_mint.to_string(),
-            pub_quote_mint.to_string(),
-            base_r,
-            quote_r,
-            0,
-        );
-        balance_update.metadata = Some(meteora_dlmm_metadata_for_pool_cache_update(&state));
-        balance_update.set_dex_readiness_in_metadata(readiness);
-        let subject = pool_subject(&pool_addr.to_string());
-        match nats.jetstream_publish(&subject, &balance_update).await {
-            Ok(true) => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    mint = %base_mint,
-                    readiness = ?readiness,
-                    "Meteora DLMM RPC refresh: published PoolCacheUpdate::BalanceUpdated to JetStream"
-                );
-                true
-            }
-            Ok(false) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    "Meteora DLMM RPC refresh: JetStream publish failed (timeout or drop)"
-                );
-                false
-            }
-            Err(e) => {
-                warn!(
-                    error = %e,
-                    request_id = %request_id,
-                    "Meteora DLMM RPC refresh: Failed to publish PoolCacheUpdate to JetStream"
-                );
-                false
-            }
-        }
-    } else {
-        false
-    };
-
-    if !jetstream_ok {
-        warn!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            "Meteora DLMM RPC refresh: MASTER cache updated but JetStream publish failed — SSOT drift risk"
-        );
-    }
-
-    Some(MeteoraDlmmRpcRefreshRowResult {
-        pool_addr,
-        readiness,
-        jetstream_published: jetstream_ok,
-    })
-}
-
-/// I-24d / Scope 21: Cold-path Meteora DLMM recovery — cache-scoped RPC refresh, JetStream SSOT, ControlResponse.
-async fn handle_ensure_meteora_dlmm_pool_state(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsureMeteoraDlmmPoolState: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pool_hint_pk = pool_address_hint.and_then(|s| Pubkey::from_str(s).ok());
-
-    if !force_refresh {
-        if let Some(pool_pk) = pool_hint_pk {
-            if ctx
-                .live_pool_cache
-                .meteora_dlmm_pool_explicitly_ready(&pool_pk)
-            {
-                info!(
-                    request_id = %request_id,
-                    base_mint = %base_mint_str,
-                    pool = %pool_pk,
-                    "EnsureMeteoraDlmmPoolState: cache hit explicit Ready (skip RPC)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Ok,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        } else if ctx
-            .live_pool_cache
-            .base_mint_has_explicit_meteora_dlmm_ready_pool(&base_mint)
-        {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureMeteoraDlmmPoolState: mint already has explicit Ready Meteora DLMM pool (skip RPC)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Ok,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool_address_hint = ?pool_address_hint,
-            "EnsureMeteoraDlmmPoolState: force_refresh — always running RPC refresh path (no cache-first short-circuit)"
-        );
-    }
-
-    let mut candidate_pools: Vec<(Pubkey, MeteoraState)> = Vec::new();
-    if let Some(pool_pk) = pool_hint_pk {
-        match ctx.live_pool_cache.get(&pool_pk) {
-            Some(CachedPoolState::Meteora(s)) => {
-                if s.token_x_mint != base_mint && s.token_y_mint != base_mint {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %pool_pk,
-                        base_mint = %base_mint_str,
-                        "EnsureMeteoraDlmmPoolState: pool_address_hint DLMM row does not list base_mint"
-                    );
-                    if let Some(ref nats) = ctx.nats {
-                        publish_control_response(
-                            nats,
-                            &ctx.run_id,
-                            request_id,
-                            ControlResponseStatus::Error,
-                            Some(pool_pk.to_string()),
-                            Some("pool hint mint mismatch".to_string()),
-                        )
-                        .await;
-                    }
-                    return;
-                }
-                candidate_pools.push((pool_pk, s));
-            }
-            Some(_) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    "EnsureMeteoraDlmmPoolState: pool_address_hint is not a Meteora DLMM row"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Error,
-                        Some(pool_pk.to_string()),
-                        Some("pool hint is not meteora dlmm in cache".to_string()),
-                    )
-                    .await;
-                }
-                return;
-            }
-            None => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    base_mint = %base_mint_str,
-                    "EnsureMeteoraDlmmPoolState: pool hint not in LivePoolCache (NotFound)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::NotFound,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        }
-    } else {
-        candidate_pools = ctx.live_pool_cache.meteora_dlmm_pools_for_mint(&base_mint);
-        if candidate_pools.is_empty() {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureMeteoraDlmmPoolState: no Meteora DLMM rows in LivePoolCache for mint (NotFound)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::NotFound,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    }
-
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        pool_address_hint = ?pool_address_hint,
-        candidates = candidate_pools.len(),
-        force_refresh = %force_refresh,
-        "EnsureMeteoraDlmmPoolState: starting cache-scoped Meteora DLMM RPC refresh"
-    );
-
-    let mut terminal_ok_pool: Option<Pubkey> = None;
-    let mut ready_jetstream_failed_pool: Option<Pubkey> = None;
-
-    for (pool_addr, state) in candidate_pools {
-        if let Some(row) = cold_path_rpc_refresh_meteora_dlmm_pool_row(
-            ctx, rpc, run_id, request_id, &base_mint, pool_addr, state,
-        )
-        .await
-        {
-            match (row.readiness, row.jetstream_published) {
-                (DexPoolReadiness::Ready, true) => {
-                    terminal_ok_pool = Some(row.pool_addr);
-                    break;
-                }
-                (DexPoolReadiness::Ready, false) => {
-                    ready_jetstream_failed_pool = Some(row.pool_addr);
-                }
-                (DexPoolReadiness::Partial, true) => {
-                    debug!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureMeteoraDlmmPoolState: Partial published to JetStream — scan remaining candidates for Ready"
-                    );
-                }
-                (DexPoolReadiness::Partial, false) => {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureMeteoraDlmmPoolState: Partial row MASTER-updated but JetStream publish failed"
-                    );
-                }
-                (DexPoolReadiness::Observed, _) => {}
-            }
-        }
-    }
-
-    if let Some(pool_pk) = terminal_ok_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                readiness = ?DexPoolReadiness::Ready,
-                jetstream = "published",
-                "EnsureMeteoraDlmmPoolState: terminal ok (Meteora DLMM Ready + JetStream publish)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Ok,
-                Some(pool_str),
-                None,
-            )
-            .await;
-        }
-        return;
-    }
-
-    if let Some(pool_pk) = ready_jetstream_failed_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            let msg = "JetStream publish failed".to_string();
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                error = %msg,
-                "EnsureMeteoraDlmmPoolState: terminal error (Ready on MASTER but JetStream publish failed)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Error,
-                Some(pool_str),
-                Some(msg),
-            )
-            .await;
-        }
-        return;
-    }
-
-    warn!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        "EnsureMeteoraDlmmPoolState: RPC refresh did not yield Meteora DLMM Ready + JetStream publish (Error)"
-    );
-    if let Some(ref nats) = ctx.nats {
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            ControlResponseStatus::Error,
-            None,
-            Some("meteora dlmm rpc refresh did not produce jetstream-ready state".to_string()),
-        )
-        .await;
-    }
-}
-
-/// Result of one cache-scoped Raydium AMM v4 RPC refresh (pool + vaults + optional Serum).
-struct RaydiumAmmRpcRefreshRowResult {
-    pool_addr: Pubkey,
-    readiness: DexPoolReadiness,
-    /// `true` when a non-`Observed` [`DexPoolReadiness`] update was published to JetStream.
-    jetstream_published: bool,
-}
-
-/// Cold-path only: refresh one Raydium AMM v4 row already present in MASTER cache — no global scan.
-/// Fetches pool account, vault balances, and Serum/OpenBook bids/asks/event_queue when `market_id`
-/// is set and serum pointers are missing (same authoritative path as FIX-29 Geyser follow-up).
-async fn cold_path_rpc_refresh_raydium_amm_pool_row(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint: &Pubkey,
-    pool_addr: Pubkey,
-    mut state: RaydiumAmmState,
-) -> Option<RaydiumAmmRpcRefreshRowResult> {
-    let raydium_program = Pubkey::from_str(RAYDIUM_AMM_V4).expect("RAYDIUM_AMM_V4 constant");
-
-    let pool_acc = match rpc.get_account_opt_retry(&pool_addr).await {
-        Ok(Some(acc)) => acc,
-        Ok(None) | Err(_) => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Raydium AMM RPC refresh: pool account fetch miss, skip"
-            );
-            return None;
-        }
-    };
-
-    let parsed_state = match parse_pool_account(&raydium_program, &pool_acc.data) {
-        Some(CachedPoolState::RaydiumAmm(s)) => s,
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                data_len = pool_acc.data.len(),
-                "Raydium AMM RPC refresh: parse pool failed, skip"
-            );
-            return None;
-        }
-    };
-
-    state.base_mint = parsed_state.base_mint;
-    state.quote_mint = parsed_state.quote_mint;
-    state.coin_vault = parsed_state.coin_vault;
-    state.pc_vault = parsed_state.pc_vault;
-    state.base_decimals = parsed_state.base_decimals;
-    state.quote_decimals = parsed_state.quote_decimals;
-    state.market_id = parsed_state.market_id;
-
-    if state.base_mint != *base_mint && state.quote_mint != *base_mint {
-        return None;
-    }
-
-    let coin_vault = state.coin_vault;
-    let pc_vault = state.pc_vault;
-    let bal_coin = match rpc.get_account_opt_retry(&coin_vault).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let bal_pc = match rpc.get_account_opt_retry(&pc_vault).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let (bc, bq) = match (bal_coin, bal_pc) {
-        (Some(a), Some(b)) => (a, b),
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Raydium AMM RPC refresh: vault balance RPC incomplete, skip"
-            );
-            return None;
-        }
-    };
-
-    state.coin_reserve = Some(bc);
-    state.pc_reserve = Some(bq);
-
-    ctx.live_pool_cache
-        .upsert(pool_addr, CachedPoolState::RaydiumAmm(state.clone()), 0);
-
-    // Serum/OpenBook static accounts (same as FIX-29 one-shot path in Geyser handler).
-    if state.market_id != Pubkey::default()
-        && (state.serum_bids.is_none()
-            || state.serum_asks.is_none()
-            || state.serum_event_queue.is_none())
-    {
-        match rpc.get_account_retry(&state.market_id).await {
-            Ok(account) => {
-                if let Some((bids_o, asks_o, eq_o, _bv, _qv)) =
-                    Raydium::parse_serum_market_accounts(&account.data)
-                {
-                    if let (Some(b), Some(a), Some(e)) = (bids_o, asks_o, eq_o) {
-                        ctx.live_pool_cache
-                            .set_raydium_serum_accounts(&pool_addr, b, a, e);
-                        ctx.raydium_serum_fetched.write().insert(pool_addr);
-                    }
-                }
-            }
-            Err(e) => {
-                debug!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    market_id = %state.market_id,
-                    error = %e,
-                    "Raydium AMM RPC refresh: serum market fetch failed (may stay Partial)"
-                );
-            }
-        }
-    }
-
-    let readiness = match ctx.live_pool_cache.get(&pool_addr).as_ref() {
-        Some(CachedPoolState::RaydiumAmm(st)) => raydium_amm_readiness_for_pool_cache_update(st),
-        _ => DexPoolReadiness::Observed,
-    };
-    ctx.live_pool_cache
-        .merge_raydium_amm_pool_readiness(pool_addr, readiness);
-
-    if readiness == DexPoolReadiness::Observed {
-        debug!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            mint = %base_mint,
-            "Raydium AMM RPC refresh: still Observed after RPC, no JetStream publish"
-        );
-        return Some(RaydiumAmmRpcRefreshRowResult {
-            pool_addr,
-            readiness,
-            jetstream_published: false,
-        });
-    }
-
-    let jetstream_ok = if let Some(ref nats) = ctx.nats {
-        if let Some((CachedPoolState::RaydiumAmm(st), pool_slot, _age_ms)) =
-            ctx.live_pool_cache.get_with_metadata(&pool_addr)
-        {
-            let mut balance_update = PoolCacheUpdate::new_balance_updated(
-                "market-data",
-                BUILD_VERSION,
-                run_id,
-                pool_addr.to_string(),
-                "raydium".to_string(),
-                st.base_mint.to_string(),
-                st.quote_mint.to_string(),
-                st.coin_reserve.unwrap_or(0),
-                st.pc_reserve.unwrap_or(0),
-                pool_slot,
-            );
-            let mut meta = std::collections::HashMap::new();
-            if st.market_id != Pubkey::default() {
-                meta.insert("market_id".to_string(), st.market_id.to_string());
-            }
-            if let (Some(bids), Some(asks), Some(eq)) =
-                (st.serum_bids, st.serum_asks, st.serum_event_queue)
-            {
-                meta.insert("serum_bids".to_string(), bids.to_string());
-                meta.insert("serum_asks".to_string(), asks.to_string());
-                meta.insert("serum_event_queue".to_string(), eq.to_string());
-            }
-            if !meta.is_empty() {
-                balance_update.metadata = Some(meta);
-            }
-            balance_update.set_dex_readiness_in_metadata(readiness);
-            let subject = pool_subject(&pool_addr.to_string());
-            match nats.jetstream_publish(&subject, &balance_update).await {
-                Ok(true) => {
-                    info!(
-                        request_id = %request_id,
-                        pool = %pool_addr,
-                        mint = %base_mint,
-                        readiness = ?readiness,
-                        "Raydium AMM RPC refresh: published PoolCacheUpdate::BalanceUpdated to JetStream"
-                    );
-                    true
-                }
-                Ok(false) => {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %pool_addr,
-                        "Raydium AMM RPC refresh: JetStream publish failed (timeout or drop)"
-                    );
-                    false
-                }
-                Err(e) => {
-                    warn!(
-                        error = %e,
-                        request_id = %request_id,
-                        "Raydium AMM RPC refresh: Failed to publish PoolCacheUpdate to JetStream"
-                    );
-                    false
-                }
-            }
-        } else {
-            false
-        }
-    } else {
-        false
-    };
-
-    if !jetstream_ok {
-        warn!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            "Raydium AMM RPC refresh: MASTER cache updated but JetStream publish failed — SSOT drift risk"
-        );
-    }
-
-    Some(RaydiumAmmRpcRefreshRowResult {
-        pool_addr,
-        readiness,
-        jetstream_published: jetstream_ok,
-    })
-}
-
-/// I-24d Scope 22: Cold-path Raydium AMM v4 recovery — cache-scoped RPC refresh, JetStream SSOT,
-/// [`ControlResponse`].
-async fn handle_ensure_raydium_amm_pool_state(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsureRaydiumAmmPoolState: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pool_hint_pk = pool_address_hint.and_then(|s| Pubkey::from_str(s).ok());
-
-    if !force_refresh {
-        if let Some(pool_pk) = pool_hint_pk {
-            if ctx
-                .live_pool_cache
-                .raydium_amm_pool_explicitly_ready(&pool_pk)
-            {
-                info!(
-                    request_id = %request_id,
-                    base_mint = %base_mint_str,
-                    pool = %pool_pk,
-                    "EnsureRaydiumAmmPoolState: cache hit explicit Ready (skip RPC)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Ok,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        } else if ctx
-            .live_pool_cache
-            .base_mint_has_explicit_raydium_amm_ready_pool(&base_mint)
-        {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureRaydiumAmmPoolState: mint already has explicit Ready Raydium AMM pool (skip RPC)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Ok,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool_address_hint = ?pool_address_hint,
-            "EnsureRaydiumAmmPoolState: force_refresh — always running RPC refresh path (no cache-first short-circuit)"
-        );
-    }
-
-    let mut candidate_pools: Vec<(Pubkey, RaydiumAmmState)> = Vec::new();
-    if let Some(pool_pk) = pool_hint_pk {
-        match ctx.live_pool_cache.get(&pool_pk) {
-            Some(CachedPoolState::RaydiumAmm(s)) => {
-                if s.base_mint != base_mint && s.quote_mint != base_mint {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %pool_pk,
-                        base_mint = %base_mint_str,
-                        "EnsureRaydiumAmmPoolState: pool_address_hint Raydium AMM row does not list base_mint"
-                    );
-                    if let Some(ref nats) = ctx.nats {
-                        publish_control_response(
-                            nats,
-                            &ctx.run_id,
-                            request_id,
-                            ControlResponseStatus::Error,
-                            Some(pool_pk.to_string()),
-                            Some("pool hint mint mismatch".to_string()),
-                        )
-                        .await;
-                    }
-                    return;
-                }
-                candidate_pools.push((pool_pk, s));
-            }
-            Some(_) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    "EnsureRaydiumAmmPoolState: pool_address_hint is not a Raydium AMM row"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Error,
-                        Some(pool_pk.to_string()),
-                        Some("pool hint is not raydium amm in cache".to_string()),
-                    )
-                    .await;
-                }
-                return;
-            }
-            None => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    base_mint = %base_mint_str,
-                    "EnsureRaydiumAmmPoolState: pool hint not in LivePoolCache (NotFound)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::NotFound,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        }
-    } else {
-        candidate_pools = ctx.live_pool_cache.raydium_amm_pools_for_mint(&base_mint);
-        if candidate_pools.is_empty() {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureRaydiumAmmPoolState: no Raydium AMM rows in LivePoolCache for mint (NotFound)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::NotFound,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    }
-
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        pool_address_hint = ?pool_address_hint,
-        candidates = candidate_pools.len(),
-        force_refresh = %force_refresh,
-        "EnsureRaydiumAmmPoolState: starting cache-scoped Raydium AMM RPC refresh"
-    );
-
-    let mut terminal_ok_pool: Option<Pubkey> = None;
-    let mut ready_jetstream_failed_pool: Option<Pubkey> = None;
-
-    for (pool_addr, state) in candidate_pools {
-        if let Some(row) = cold_path_rpc_refresh_raydium_amm_pool_row(
-            ctx, rpc, run_id, request_id, &base_mint, pool_addr, state,
-        )
-        .await
-        {
-            match (row.readiness, row.jetstream_published) {
-                (DexPoolReadiness::Ready, true) => {
-                    terminal_ok_pool = Some(row.pool_addr);
-                    break;
-                }
-                (DexPoolReadiness::Ready, false) => {
-                    ready_jetstream_failed_pool = Some(row.pool_addr);
-                }
-                (DexPoolReadiness::Partial, true) => {
-                    debug!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureRaydiumAmmPoolState: Partial published to JetStream — scan remaining candidates for Ready"
-                    );
-                }
-                (DexPoolReadiness::Partial, false) => {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureRaydiumAmmPoolState: Partial row MASTER-updated but JetStream publish failed"
-                    );
-                }
-                (DexPoolReadiness::Observed, _) => {}
-            }
-        }
-    }
-
-    if let Some(pool_pk) = terminal_ok_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                readiness = ?DexPoolReadiness::Ready,
-                jetstream = "published",
-                "EnsureRaydiumAmmPoolState: terminal ok (Raydium AMM Ready + JetStream publish)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Ok,
-                Some(pool_str),
-                None,
-            )
-            .await;
-        }
-        return;
-    }
-
-    if let Some(pool_pk) = ready_jetstream_failed_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            let msg = "JetStream publish failed".to_string();
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                error = %msg,
-                "EnsureRaydiumAmmPoolState: terminal error (Raydium AMM Ready on MASTER but JetStream publish failed)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Error,
-                Some(pool_str),
-                Some(msg),
-            )
-            .await;
-        }
-        return;
-    }
-
-    warn!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        "EnsureRaydiumAmmPoolState: RPC refresh did not yield Raydium AMM Ready + JetStream publish (Error)"
-    );
-    if let Some(ref nats) = ctx.nats {
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            ControlResponseStatus::Error,
-            None,
-            Some("raydium amm rpc refresh did not produce jetstream-ready state".to_string()),
-        )
-        .await;
-    }
-}
-
-/// Result of one cache-scoped Raydium CPMM RPC refresh (pool + vault balances).
-struct RaydiumCpmmRpcRefreshRowResult {
-    pool_addr: Pubkey,
-    readiness: DexPoolReadiness,
-    /// `true` when a non-`Observed` [`DexPoolReadiness`] update was published to JetStream.
-    jetstream_published: bool,
-}
-
-/// Cold-path only: refresh one Raydium CPMM row already present in MASTER cache — no global scan.
-async fn cold_path_rpc_refresh_raydium_cpmm_pool_row(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint: &Pubkey,
-    pool_addr: Pubkey,
-    mut state: RaydiumCpmmState,
-) -> Option<RaydiumCpmmRpcRefreshRowResult> {
-    let raydium_cpmm_program = Pubkey::from_str(RAYDIUM_CPMM).expect("RAYDIUM_CPMM constant");
-    let sol = Pubkey::from_str(NATIVE_SOL_MINT).unwrap_or_default();
-
-    let pool_acc = match rpc.get_account_opt_retry(&pool_addr).await {
-        Ok(Some(acc)) => acc,
-        Ok(None) | Err(_) => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Raydium CPMM RPC refresh: pool account fetch miss, skip"
-            );
-            return None;
-        }
-    };
-
-    let parsed_state = match parse_pool_account(&raydium_cpmm_program, &pool_acc.data) {
-        Some(CachedPoolState::RaydiumCpmm(s)) => s,
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                data_len = pool_acc.data.len(),
-                "Raydium CPMM RPC refresh: parse pool failed, skip"
-            );
-            return None;
-        }
-    };
-
-    state.token_0_mint = parsed_state.token_0_mint;
-    state.token_1_mint = parsed_state.token_1_mint;
-    state.token_0_vault = parsed_state.token_0_vault;
-    state.token_1_vault = parsed_state.token_1_vault;
-
-    if state.token_0_mint != *base_mint && state.token_1_mint != *base_mint {
-        return None;
-    }
-
-    let vault0 = state.token_0_vault;
-    let vault1 = state.token_1_vault;
-    let bal0 = match rpc.get_account_opt_retry(&vault0).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let bal1 = match rpc.get_account_opt_retry(&vault1).await {
-        Ok(Some(acc)) => try_parse_token_account_balance(&acc.data),
-        _ => None,
-    };
-    let (b0, b1) = match (bal0, bal1) {
-        (Some(a), Some(b)) => (a, b),
-        _ => {
-            debug!(
-                request_id = %request_id,
-                pool = %pool_addr,
-                "Raydium CPMM RPC refresh: vault balance RPC incomplete, skip"
-            );
-            return None;
-        }
-    };
-
-    state.reserve_0 = Some(b0);
-    state.reserve_1 = Some(b1);
-
-    ctx.live_pool_cache
-        .upsert(pool_addr, CachedPoolState::RaydiumCpmm(state.clone()), 0);
-
-    let readiness = raydium_cpmm_readiness_for_pool_cache_update(&state);
-    ctx.live_pool_cache
-        .merge_raydium_cpmm_pool_readiness(pool_addr, readiness);
-
-    if readiness == DexPoolReadiness::Observed {
-        debug!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            mint = %base_mint,
-            "Raydium CPMM RPC refresh: still Observed after RPC, no JetStream publish"
-        );
-        return Some(RaydiumCpmmRpcRefreshRowResult {
-            pool_addr,
-            readiness,
-            jetstream_published: false,
-        });
-    }
-
-    let (pub_base_mint, pub_quote_mint, base_r, quote_r) = if state.token_1_mint == sol {
-        (state.token_0_mint, state.token_1_mint, b0, b1)
-    } else if state.token_0_mint == sol {
-        (state.token_1_mint, state.token_0_mint, b1, b0)
-    } else {
-        (state.token_0_mint, state.token_1_mint, b0, b1)
-    };
-
-    let jetstream_ok = if let Some(ref nats) = ctx.nats {
-        let mut balance_update = PoolCacheUpdate::new_balance_updated(
-            "market-data",
-            BUILD_VERSION,
-            run_id,
-            pool_addr.to_string(),
-            "raydium_cpmm".to_string(),
-            pub_base_mint.to_string(),
-            pub_quote_mint.to_string(),
-            base_r,
-            quote_r,
-            0,
-        );
-        let mut meta = std::collections::HashMap::new();
-        meta.insert(
-            POOL_CACHE_UPDATE_RAYDIUM_CPMM_VAULTS_KEY.to_string(),
-            raydium_cpmm_vaults_for_pool_cache_update(&state),
-        );
-        balance_update.metadata = Some(meta);
-        balance_update.set_dex_readiness_in_metadata(readiness);
-        let subject = pool_subject(&pool_addr.to_string());
-        match nats.jetstream_publish(&subject, &balance_update).await {
-            Ok(true) => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    mint = %base_mint,
-                    readiness = ?readiness,
-                    "Raydium CPMM RPC refresh: published PoolCacheUpdate::BalanceUpdated to JetStream"
-                );
-                true
-            }
-            Ok(false) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_addr,
-                    "Raydium CPMM RPC refresh: JetStream publish failed (timeout or drop)"
-                );
-                false
-            }
-            Err(e) => {
-                warn!(
-                    error = %e,
-                    request_id = %request_id,
-                    "Raydium CPMM RPC refresh: Failed to publish PoolCacheUpdate to JetStream"
-                );
-                false
-            }
-        }
-    } else {
-        false
-    };
-
-    if !jetstream_ok {
-        warn!(
-            request_id = %request_id,
-            pool = %pool_addr,
-            "Raydium CPMM RPC refresh: MASTER cache updated but JetStream publish failed — SSOT drift risk"
-        );
-    }
-
-    Some(RaydiumCpmmRpcRefreshRowResult {
-        pool_addr,
-        readiness,
-        jetstream_published: jetstream_ok,
-    })
-}
-
-/// I-24d: Cold-path Raydium CPMM recovery — cache-scoped RPC refresh, JetStream SSOT,
-/// [`ControlResponse`].
-async fn handle_ensure_raydium_cpmm_pool_state(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsureRaydiumCpmmPoolState: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pool_hint_pk = pool_address_hint.and_then(|s| Pubkey::from_str(s).ok());
-
-    if !force_refresh {
-        if let Some(pool_pk) = pool_hint_pk {
-            if ctx
-                .live_pool_cache
-                .raydium_cpmm_pool_explicitly_ready(&pool_pk)
-            {
-                info!(
-                    request_id = %request_id,
-                    base_mint = %base_mint_str,
-                    pool = %pool_pk,
-                    "EnsureRaydiumCpmmPoolState: cache hit explicit Ready (skip RPC)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Ok,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        } else if ctx
-            .live_pool_cache
-            .base_mint_has_explicit_raydium_cpmm_ready_pool(&base_mint)
-        {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureRaydiumCpmmPoolState: mint already has explicit Ready Raydium CPMM pool (skip RPC)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Ok,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool_address_hint = ?pool_address_hint,
-            "EnsureRaydiumCpmmPoolState: force_refresh — always running RPC refresh path (no cache-first short-circuit)"
-        );
-    }
-
-    let mut candidate_pools: Vec<(Pubkey, RaydiumCpmmState)> = Vec::new();
-    if let Some(pool_pk) = pool_hint_pk {
-        match ctx.live_pool_cache.get(&pool_pk) {
-            Some(CachedPoolState::RaydiumCpmm(s)) => {
-                if s.token_0_mint != base_mint && s.token_1_mint != base_mint {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %pool_pk,
-                        base_mint = %base_mint_str,
-                        "EnsureRaydiumCpmmPoolState: pool_address_hint Raydium CPMM row does not list base_mint"
-                    );
-                    if let Some(ref nats) = ctx.nats {
-                        publish_control_response(
-                            nats,
-                            &ctx.run_id,
-                            request_id,
-                            ControlResponseStatus::Error,
-                            Some(pool_pk.to_string()),
-                            Some("pool hint mint mismatch".to_string()),
-                        )
-                        .await;
-                    }
-                    return;
-                }
-                candidate_pools.push((pool_pk, s));
-            }
-            Some(_) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    "EnsureRaydiumCpmmPoolState: pool_address_hint is not a Raydium CPMM row"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Error,
-                        Some(pool_pk.to_string()),
-                        Some("pool hint is not raydium cpmm in cache".to_string()),
-                    )
-                    .await;
-                }
-                return;
-            }
-            None => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    base_mint = %base_mint_str,
-                    "EnsureRaydiumCpmmPoolState: pool hint not in LivePoolCache (NotFound)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::NotFound,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        }
-    } else {
-        candidate_pools = ctx.live_pool_cache.raydium_cpmm_pools_for_mint(&base_mint);
-        if candidate_pools.is_empty() {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureRaydiumCpmmPoolState: no Raydium CPMM rows in LivePoolCache for mint (NotFound)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::NotFound,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    }
-
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        pool_address_hint = ?pool_address_hint,
-        candidates = candidate_pools.len(),
-        force_refresh = %force_refresh,
-        "EnsureRaydiumCpmmPoolState: starting cache-scoped Raydium CPMM RPC refresh"
-    );
-
-    let mut terminal_ok_pool: Option<Pubkey> = None;
-    let mut ready_jetstream_failed_pool: Option<Pubkey> = None;
-
-    for (pool_addr, state) in candidate_pools {
-        if let Some(row) = cold_path_rpc_refresh_raydium_cpmm_pool_row(
-            ctx, rpc, run_id, request_id, &base_mint, pool_addr, state,
-        )
-        .await
-        {
-            match (row.readiness, row.jetstream_published) {
-                (DexPoolReadiness::Ready, true) => {
-                    terminal_ok_pool = Some(row.pool_addr);
-                    break;
-                }
-                (DexPoolReadiness::Ready, false) => {
-                    ready_jetstream_failed_pool = Some(row.pool_addr);
-                }
-                (DexPoolReadiness::Partial, true) => {
-                    debug!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureRaydiumCpmmPoolState: Partial published to JetStream — scan remaining candidates for Ready"
-                    );
-                }
-                (DexPoolReadiness::Partial, false) => {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureRaydiumCpmmPoolState: Partial row MASTER-updated but JetStream publish failed"
-                    );
-                }
-                (DexPoolReadiness::Observed, _) => {}
-            }
-        }
-    }
-
-    if let Some(pool_pk) = terminal_ok_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                readiness = ?DexPoolReadiness::Ready,
-                jetstream = "published",
-                "EnsureRaydiumCpmmPoolState: terminal ok (Raydium CPMM Ready + JetStream publish)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Ok,
-                Some(pool_str),
-                None,
-            )
-            .await;
-        }
-        return;
-    }
-
-    if let Some(pool_pk) = ready_jetstream_failed_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            let msg = "JetStream publish failed".to_string();
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                error = %msg,
-                "EnsureRaydiumCpmmPoolState: terminal error (Raydium CPMM Ready on MASTER but JetStream publish failed)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Error,
-                Some(pool_str),
-                Some(msg),
-            )
-            .await;
-        }
-        return;
-    }
-
-    warn!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        "EnsureRaydiumCpmmPoolState: RPC refresh did not yield Raydium CPMM Ready + JetStream publish (Error)"
-    );
-    if let Some(ref nats) = ctx.nats {
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            ControlResponseStatus::Error,
-            None,
-            Some("raydium cpmm rpc refresh did not produce jetstream-ready state".to_string()),
-        )
-        .await;
-    }
-}
-
-/// I-24d: Cold-path Meteora CPMM recovery — cache-scoped RPC refresh, JetStream SSOT,
-/// [`ControlResponse`].
-async fn handle_ensure_meteora_cpmm_pool_state(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsureMeteoraCpmmPoolState: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pool_hint_pk = pool_address_hint.and_then(|s| Pubkey::from_str(s).ok());
-
-    if !force_refresh {
-        if let Some(pool_pk) = pool_hint_pk {
-            if ctx
-                .live_pool_cache
-                .meteora_cpmm_pool_explicitly_ready(&pool_pk)
-            {
-                info!(
-                    request_id = %request_id,
-                    base_mint = %base_mint_str,
-                    pool = %pool_pk,
-                    "EnsureMeteoraCpmmPoolState: cache hit explicit Ready (skip RPC)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Ok,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        } else if ctx
-            .live_pool_cache
-            .base_mint_has_explicit_meteora_cpmm_ready_pool(&base_mint)
-        {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureMeteoraCpmmPoolState: mint already has explicit Ready Meteora CPMM pool (skip RPC)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Ok,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool_address_hint = ?pool_address_hint,
-            "EnsureMeteoraCpmmPoolState: force_refresh — always running RPC refresh path (no cache-first short-circuit)"
-        );
-    }
-
-    let mut candidate_pools: Vec<(Pubkey, MeteoraCpmmState)> = Vec::new();
-    if let Some(pool_pk) = pool_hint_pk {
-        match ctx.live_pool_cache.get(&pool_pk) {
-            Some(CachedPoolState::MeteoraCpmm(s)) => {
-                if s.token_0_mint != base_mint && s.token_1_mint != base_mint {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %pool_pk,
-                        base_mint = %base_mint_str,
-                        "EnsureMeteoraCpmmPoolState: pool_address_hint Meteora CPMM row does not list base_mint"
-                    );
-                    if let Some(ref nats) = ctx.nats {
-                        publish_control_response(
-                            nats,
-                            &ctx.run_id,
-                            request_id,
-                            ControlResponseStatus::Error,
-                            Some(pool_pk.to_string()),
-                            Some("pool hint mint mismatch".to_string()),
-                        )
-                        .await;
-                    }
-                    return;
-                }
-                candidate_pools.push((pool_pk, s));
-            }
-            Some(_) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    "EnsureMeteoraCpmmPoolState: pool_address_hint is not a Meteora CPMM row"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Error,
-                        Some(pool_pk.to_string()),
-                        Some("pool hint is not meteora cpmm in cache".to_string()),
-                    )
-                    .await;
-                }
-                return;
-            }
-            None => {
-                info!(
-                    request_id = %request_id,
-                    pool = %pool_pk,
-                    base_mint = %base_mint_str,
-                    "EnsureMeteoraCpmmPoolState: pool hint not in LivePoolCache (NotFound)"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::NotFound,
-                        Some(pool_pk.to_string()),
-                        None,
-                    )
-                    .await;
-                }
-                return;
-            }
-        }
-    } else {
-        candidate_pools = ctx.live_pool_cache.meteora_cpmm_pools_for_mint(&base_mint);
-        if candidate_pools.is_empty() {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                "EnsureMeteoraCpmmPoolState: no Meteora CPMM rows in LivePoolCache for mint (NotFound)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::NotFound,
-                    None,
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    }
-
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        pool_address_hint = ?pool_address_hint,
-        candidates = candidate_pools.len(),
-        force_refresh = %force_refresh,
-        "EnsureMeteoraCpmmPoolState: starting cache-scoped Meteora CPMM RPC refresh"
-    );
-
-    let mut terminal_ok_pool: Option<Pubkey> = None;
-    let mut ready_jetstream_failed_pool: Option<Pubkey> = None;
-
-    for (pool_addr, state) in candidate_pools {
-        if let Some(row) = cold_path_rpc_refresh_meteora_cpmm_pool_row(
-            ctx, rpc, run_id, request_id, &base_mint, pool_addr, state,
-        )
-        .await
-        {
-            match (row.readiness, row.jetstream_published) {
-                (DexPoolReadiness::Ready, true) => {
-                    terminal_ok_pool = Some(row.pool_addr);
-                    break;
-                }
-                (DexPoolReadiness::Ready, false) => {
-                    ready_jetstream_failed_pool = Some(row.pool_addr);
-                }
-                (DexPoolReadiness::Partial, true) => {
-                    debug!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureMeteoraCpmmPoolState: Partial published to JetStream — scan remaining candidates for Ready"
-                    );
-                }
-                (DexPoolReadiness::Partial, false) => {
-                    warn!(
-                        request_id = %request_id,
-                        pool = %row.pool_addr,
-                        "EnsureMeteoraCpmmPoolState: Partial row MASTER-updated but JetStream publish failed"
-                    );
-                }
-                (DexPoolReadiness::Observed, _) => {}
-            }
-        }
-    }
-
-    if let Some(pool_pk) = terminal_ok_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                readiness = ?DexPoolReadiness::Ready,
-                jetstream = "published",
-                "EnsureMeteoraCpmmPoolState: terminal ok (Meteora CPMM Ready + JetStream publish)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Ok,
-                Some(pool_str),
-                None,
-            )
-            .await;
-        }
-        return;
-    }
-
-    if let Some(pool_pk) = ready_jetstream_failed_pool {
-        if let Some(ref nats) = ctx.nats {
-            let pool_str = pool_pk.to_string();
-            let msg = "JetStream publish failed".to_string();
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_str,
-                error = %msg,
-                "EnsureMeteoraCpmmPoolState: terminal error (Meteora CPMM Ready on MASTER but JetStream publish failed)"
-            );
-            publish_control_response(
-                nats,
-                &ctx.run_id,
-                request_id,
-                ControlResponseStatus::Error,
-                Some(pool_str),
-                Some(msg),
-            )
-            .await;
-        }
-        return;
-    }
-
-    warn!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        "EnsureMeteoraCpmmPoolState: RPC refresh did not yield Meteora CPMM Ready + JetStream publish (Error)"
-    );
-    if let Some(ref nats) = ctx.nats {
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            ControlResponseStatus::Error,
-            None,
-            Some("meteora cpmm rpc refresh did not produce jetstream-ready state".to_string()),
-        )
-        .await;
-    }
-}
-
-/// Cold-path only: Meteora DLMM promotion for one wallet mint. Uses **only** pools already present
-/// in [`LivePoolCache`] (from Geyser); no `getProgramAccounts` / global scan. RPC: per-pool
-/// `get_account` (LB pair) + per-reserve-vault balance reads — bounded by matching cache rows.
-///
-/// Publishes [`PoolCacheUpdate`] with [`meteora_dlmm_metadata_for_pool_cache_update`] and
-/// [`meteora_dlmm_readiness_for_pool_cache_update`] like the Geyser path.
-///
-/// **Slot:** `geyser_slot = 0` on [`LivePoolCache::upsert`] and [`PoolCacheUpdate::new_balance_updated`],
 /// same as other bounded wallet-bootstrap verify handlers.
 async fn handle_wallet_bootstrap_meteora_dlmm_verify_for_mint(
     ctx: &MarketDataContext,
@@ -7402,7 +4746,7 @@ async fn handle_wallet_bootstrap_meteora_dlmm_verify_for_mint(
             break;
         }
 
-        let _ = cold_path_rpc_refresh_meteora_dlmm_pool_row(
+        cold_path_rpc_refresh_meteora_dlmm_pool_row(
             ctx, rpc, run_id, request_id, base_mint, pool_addr, state,
         )
         .await;
@@ -8821,949 +6165,6 @@ async fn main() -> Result<()> {
     info!(run_id = %run_id, "market-data shutdown complete");
 
     Ok(())
-}
-
-/// Cold Path: merge existing PumpAmm reserves with authoritative vault RPC reads for I-24d Ensure.
-///
-/// - If the cache already has **both** reserves present and strictly positive, returns `Ok` with
-///   those values (skip vault RPC).
-/// - Otherwise vault RPC **must** succeed. On RPC failure there is **no** silent fallback to
-///   `0/0` — that would let the handler publish a false “success” SSOT (I-24d / I-12).
-async fn resolve_pump_amm_reserves_for_ensure_discovery(
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_str: &str,
-    existing: Option<&CachedPoolState>,
-    pool_accounts: &[Pubkey],
-    dex: &ironcrab::solana::dex::pumpfun_amm::PumpFunAmmDex,
-    force_refresh: bool,
-) -> Result<(u64, u64), String> {
-    if !force_refresh {
-        if let Some(CachedPoolState::PumpAmm(s)) = existing {
-            if let (Some(br), Some(qr)) = (s.base_reserve, s.quote_reserve) {
-                if br > 0 && qr > 0 {
-                    info!(
-                        request_id = %request_id,
-                        base_mint = %base_mint_str,
-                        pool = %pool_address_str,
-                        base_reserve = br,
-                        quote_reserve = qr,
-                        "EnsurePumpAmmPoolAccounts: using existing non-degenerate reserves (skip vault RPC)"
-                    );
-                    return Ok((br, qr));
-                }
-            }
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool = %pool_address_str,
-            "EnsurePumpAmmPoolAccounts: force_refresh — always hydrating vault reserves via RPC (no cache-first reserves)"
-        );
-    }
-
-    match dex.fetch_pump_amm_vault_reserves(pool_accounts).await {
-        Ok((b, q)) => {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_address_str,
-                base_reserve = b,
-                quote_reserve = q,
-                "EnsurePumpAmmPoolAccounts: hydrated vault reserves via RPC (Cold Path)"
-            );
-            Ok((b, q))
-        }
-        Err(e) => {
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %pool_address_str,
-                error = %e,
-                "EnsurePumpAmmPoolAccounts: vault reserve RPC failed (no non-degenerate cached reserves)"
-            );
-            Err(format!("vault reserve RPC failed: {e}"))
-        }
-    }
-}
-
-/// Scope 44: one structured `info!` for cold-path PumpSwap v14 provenance (SIM 6023 forensics).
-fn log_pump_amm_scope44_pool_accounts_diag(
-    request_id: &str,
-    base_mint_str: &str,
-    force_refresh: bool,
-    diag: &PumpAmmPoolAccountsDiagnostic,
-    v14: &[Pubkey],
-) {
-    let ref_sig = diag.reference_swap_signature.as_deref().unwrap_or("");
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        scope = "44",
-        pump_amm_diag_source = diag.source,
-        pool_market = %diag.pool_market,
-        diag_force_refresh = diag.force_refresh,
-        request_force_refresh = %force_refresh,
-        ref_swap_sig = %ref_sig,
-        v14_csv = %PumpAmmPoolAccountsDiagnostic::format_v14_csv(v14),
-        pfr = %v14.get(6).copied().unwrap_or_default(),
-        pfr_field = ?diag.protocol_fee_recipient.resolution,
-        pfr_tag = diag.protocol_fee_recipient.tag,
-        pfr_ta = %v14.get(7).copied().unwrap_or_default(),
-        pfr_ta_field = ?diag.protocol_fee_recipient_ta.resolution,
-        pfr_ta_tag = diag.protocol_fee_recipient_ta.tag,
-        cc_vault_ata = %v14.get(9).copied().unwrap_or_default(),
-        cc_vault_ata_field = ?diag.coin_creator_vault_ata.resolution,
-        cc_vault_ata_tag = diag.coin_creator_vault_ata.tag,
-        cc_auth = %v14.get(10).copied().unwrap_or_default(),
-        cc_auth_field = ?diag.coin_creator_vault_authority.resolution,
-        cc_auth_tag = diag.coin_creator_vault_authority.tag,
-        gva = %v14.get(11).copied().unwrap_or_default(),
-        gva_field = ?diag.global_volume_accumulator.resolution,
-        gva_tag = diag.global_volume_accumulator.tag,
-        fee_cfg_cached = %v14.get(12).copied().unwrap_or_default(),
-        fee_cfg_field = ?diag.fee_config.resolution,
-        fee_cfg_tag = diag.fee_config.tag,
-        fee_prog_cached = %v14.get(13).copied().unwrap_or_default(),
-        fee_prog_field = ?diag.fee_program.resolution,
-        fee_prog_tag = diag.fee_program.tag,
-        "I-24d Scope44: pump_amm v14 diagnostic — ref_sig=successful swap TX for manual diff; SELL ix uses global fee_config + fee_program (5PH… / pfee…); v14[12]/[13] may differ (cache row)"
-    );
-}
-
-/// Scope 49: single structured line before ControlResponse=Error when force_refresh SELL layout stays unresolved.
-fn log_pump_amm_scope49_force_refresh_sell_layout_failure(
-    request_id: &str,
-    pool_address_str: &str,
-    base_mint_str: &str,
-    diag: Option<&ironcrab::solana::dex::pumpfun_amm::PumpAmmForceRefreshSellLayoutDiag>,
-) {
-    let Some(d) = diag else {
-        warn!(
-            request_id = %request_id,
-            pool = %pool_address_str,
-            base_mint = %base_mint_str,
-            sell_layout_failure_class = "force_refresh_diag_missing",
-            "I-24d Scope49: sell_layout_ready=false after force_refresh (no structured diag from dex — check pumpfun_amm version)"
-        );
-        return;
-    };
-    let ext = d.last_external.as_ref();
-    warn!(
-        request_id = %request_id,
-        pool = %pool_address_str,
-        base_mint = %base_mint_str,
-        scope = "49",
-        local_history_empty = d.local_history_empty,
-        local_observation_failed = d.local_observation_failed,
-        local_history_probe = %d.local_history_probe.as_log_str(),
-        external_attempted = d.external_attempted,
-        external_sig_limit = ?d.external_sig_limit,
-        external_max_tx_fetches = ?d.external_max_tx_fetches,
-        external_max_get_transaction_calls = ?d.external_max_get_transaction_calls,
-        termination_reason = %d.termination_reason.as_log_str(),
-        ext_elapsed_ms = ext.map(|s| s.elapsed_total_ms),
-        ext_get_signatures_calls = ext.map(|s| s.get_signatures_calls),
-        ext_get_transaction_calls = ext.map(|s| s.get_transaction_calls),
-        ext_signatures_returned = ext.map(|s| s.signatures_returned_last),
-        ext_transactions_fetched = ext.map(|s| s.transactions_fetched),
-        ext_pump_amm_ix_seen = ext.map(|s| s.pump_amm_instructions_seen),
-        ext_pump_amm_sell_discriminator_seen = ext.map(|s| s.pump_amm_sell_discriminator_seen),
-        ext_sell_candidates_seen = ext.map(|s| s.sell_candidates_seen),
-        ext_provider_status = ext.map(|s| s.provider_status_last.as_log_str()),
-        ext_termination = ext.map(|s| s.termination_reason.as_log_str()),
-        "I-24d Scope49: sell_layout_ready=false after force_refresh — classification for supervisor (timeout vs 429 vs budget vs empty)"
-    );
-}
-
-/// PR235: true when md-state queue is saturated enough to defer cold-path discovery RPC.
-fn market_data_md_state_queue_pressure_defer_discovery() -> bool {
-    let depth =
-        ironcrab::metrics::MARKET_DATA_GEYSER_TRACKING_QUEUE_DEPTH.load(Ordering::Relaxed) as f64;
-    depth
-        >= (MARKET_DATA_GEYSER_TRACKING_QUEUE_CAP as f64)
-            * MARKET_DATA_MD_STATE_DISCOVERY_DEFER_QUEUE_FRAC
-}
-
-async fn defer_discovery_if_md_state_pressure(ctx: &MarketDataContext, request_id: &str) -> bool {
-    if !market_data_md_state_queue_pressure_defer_discovery() {
-        return false;
-    }
-    inc_market_data_discovery_deferred_md_state_pressure_total();
-    let depth = ironcrab::metrics::MARKET_DATA_GEYSER_TRACKING_QUEUE_DEPTH.load(Ordering::Relaxed);
-    warn!(
-        request_id = %request_id,
-        queue_depth = depth,
-        queue_cap = MARKET_DATA_GEYSER_TRACKING_QUEUE_CAP,
-        "PR235: deferring cold-path discovery — md-state queue under pressure"
-    );
-    if let Some(ref nats) = ctx.nats {
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            ControlResponseStatus::Busy,
-            None,
-            Some("md_state_queue_pressure".into()),
-        )
-        .await;
-    }
-    true
-}
-
-/// I-24d: Handle EnsurePumpAmmPoolAccounts Discovery Request.
-///
-/// Performs RPC-based discovery (Cold Path), updates MASTER cache, publishes
-/// JetStream PoolCacheUpdate (SSOT), and ControlResponse (correlation only).
-/// Uses shared PumpFunAmmDex for dedupe/storm protection.
-/// When pool_address_hint is provided, uses fast getAccount path (<1s) instead of slow getProgramAccounts.
-async fn handle_ensure_pump_amm_pool_accounts(
-    ctx: &MarketDataContext,
-    dex: &ironcrab::solana::dex::pumpfun_amm::PumpFunAmmDex,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    if defer_discovery_if_md_state_pressure(ctx, request_id).await {
-        return;
-    }
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsurePumpAmmPoolAccounts: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pool_hint = pool_address_hint.and_then(|s| Pubkey::from_str(s).ok());
-    info!(
-        request_id = %request_id,
-        base_mint = %base_mint_str,
-        pool_address_hint = ?pool_address_hint,
-        has_pool_hint = pool_hint.is_some(),
-        force_refresh = %force_refresh,
-        "I-24d Discovery: EnsurePumpAmmPoolAccounts start (Geyser cache check then RPC fast path if hint)"
-    );
-    match dex
-        .pool_accounts_v1_for_base_mint_with_hint_diagnostic(base_mint, pool_hint, force_refresh)
-        .await
-    {
-        Ok(Some(wrapped)) if wrapped.accounts.len() >= 14 => {
-            let sell_cashback_remaining = wrapped.sell_requires_cashback_remaining;
-            let sell_cashback_third = wrapped.sell_cashback_third_meta;
-            let sell_tail_0 = wrapped.sell_extended_tail_0;
-            let sell_tail_1 = wrapped.sell_extended_tail_1;
-            let sell_fee_tail_0 = wrapped.sell_extended_fee_tail_0;
-            let sell_fee_tail_1 = wrapped.sell_extended_fee_tail_1;
-            let sell_requires_pre_fee_metas = wrapped.sell_requires_pre_fee_metas;
-            let sell_pre_fee_meta_1 = wrapped.sell_pre_fee_meta_1;
-            let sell_layout_ready_from_refresh = wrapped.sell_layout_ready;
-            let pump_amm_force_refresh_sell_diag = wrapped.force_refresh_sell_layout_diag;
-            let accounts = wrapped.accounts;
-            let diag = wrapped.diagnostic;
-            log_pump_amm_scope44_pool_accounts_diag(
-                request_id,
-                base_mint_str,
-                force_refresh,
-                &diag,
-                &accounts,
-            );
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                outcome = "ok",
-                rpc_global_scan = "no",
-                "I-24d Discovery: pump_amm pool_accounts resolved (terminal)"
-            );
-            let pool_address = accounts[0];
-            let pool_address_str = pool_address.to_string();
-            let base_mint_str = base_mint.to_string();
-            let quote_mint = accounts[3];
-            let quote_mint_str = quote_mint.to_string();
-
-            let existing = ctx.live_pool_cache.get(&pool_address);
-            let creator_opt = existing.as_ref().and_then(|st| match st {
-                CachedPoolState::PumpAmm(s) => s.creator,
-                _ => None,
-            });
-            let (ext_flag_cache, ext_third_cache, ext_t0_cache, ext_t1_cache) = ctx
-                .live_pool_cache
-                .pump_amm_sell_extended_layout(&pool_address);
-            let (ext_fee_t0_cache, ext_fee_t1_cache) = ctx
-                .live_pool_cache
-                .pump_amm_sell_fee_tail_layout(&pool_address);
-            let ext_requires_fee_cache = ctx
-                .live_pool_cache
-                .pump_amm_sell_requires_fee_tail(&pool_address);
-            let ext_requires_pre_fee_cache = ctx
-                .live_pool_cache
-                .pump_amm_sell_requires_pre_fee_metas(&pool_address);
-            let ext_pre_fee_meta_1_cache = ctx
-                .live_pool_cache
-                .pump_amm_sell_pre_fee_meta_1(&pool_address);
-            let refresh_requires_fee_tail = sell_fee_tail_0.is_some() && sell_fee_tail_1.is_some();
-            let (
-                sell_flag_merged,
-                third_merged,
-                tail0_merged,
-                tail1_merged,
-                fee_tail0_merged,
-                fee_tail1_merged,
-                requires_fee_tail_merged,
-                requires_pre_fee_metas_merged,
-                pre_fee_meta_1_merged,
-                sell_layout_ready,
-                dex_readiness,
-            ) = pump_amm_sell_layout_state_for_ensure_publish(
-                force_refresh,
-                ext_flag_cache,
-                ext_third_cache,
-                ext_t0_cache,
-                ext_t1_cache,
-                ext_fee_t0_cache,
-                ext_fee_t1_cache,
-                ext_requires_fee_cache,
-                ext_requires_pre_fee_cache,
-                ext_pre_fee_meta_1_cache,
-                sell_cashback_remaining,
-                sell_cashback_third,
-                sell_tail_0,
-                sell_tail_1,
-                sell_fee_tail_0,
-                sell_fee_tail_1,
-                refresh_requires_fee_tail,
-                sell_requires_pre_fee_metas,
-                sell_pre_fee_meta_1,
-                sell_layout_ready_from_refresh,
-            );
-
-            let (base_reserve, quote_reserve) =
-                match resolve_pump_amm_reserves_for_ensure_discovery(
-                    request_id,
-                    &base_mint_str,
-                    &pool_address_str,
-                    existing.as_ref(),
-                    &accounts,
-                    dex,
-                    force_refresh,
-                )
-                .await
-                {
-                    Ok(pair) => pair,
-                    Err(msg) => {
-                        warn!(
-                            request_id = %request_id,
-                            base_mint = %base_mint_str,
-                            pool = %pool_address_str,
-                            error = %msg,
-                            "I-24d Discovery: terminal outcome error (reserve hydration required)"
-                        );
-                        if let Some(ref nats) = ctx.nats {
-                            publish_control_response(
-                                nats,
-                                &ctx.run_id,
-                                request_id,
-                                ControlResponseStatus::Error,
-                                Some(pool_address_str),
-                                Some(msg),
-                            )
-                            .await;
-                        }
-                        return;
-                    }
-                };
-
-            let state = CachedPoolState::PumpAmm(PumpAmmState {
-                base_mint,
-                quote_mint,
-                pool_base_token_account: accounts[4],
-                pool_quote_token_account: accounts[5],
-                base_reserve: Some(base_reserve),
-                quote_reserve: Some(quote_reserve),
-                pool_accounts: accounts.clone(),
-                creator: creator_opt,
-            });
-            ctx.live_pool_cache.upsert(pool_address, state, 0);
-            if force_refresh {
-                ctx.live_pool_cache.set_pump_amm_sell_layout_authoritative(
-                    &pool_address,
-                    sell_flag_merged,
-                    third_merged,
-                    tail0_merged,
-                    tail1_merged,
-                    fee_tail0_merged,
-                    fee_tail1_merged,
-                    requires_fee_tail_merged,
-                    requires_pre_fee_metas_merged,
-                    pre_fee_meta_1_merged,
-                );
-            } else {
-                ctx.live_pool_cache.merge_pump_amm_sell_extended_layout(
-                    &pool_address,
-                    sell_flag_merged,
-                    third_merged,
-                    tail0_merged,
-                    tail1_merged,
-                    fee_tail0_merged,
-                    fee_tail1_merged,
-                    requires_fee_tail_merged,
-                    requires_pre_fee_metas_merged,
-                    pre_fee_meta_1_merged,
-                );
-            }
-            ctx.live_pool_cache
-                .set_pump_amm_sell_layout_ready(&pool_address, sell_layout_ready);
-            ctx.live_pool_cache
-                .set_pump_amm_pool_accounts_readiness_authoritative(pool_address, dex_readiness);
-
-            let layout_generation = if force_refresh {
-                ctx.live_pool_cache
-                    .bump_pump_amm_layout_generation(&pool_address)
-            } else {
-                ctx.live_pool_cache
-                    .pump_amm_layout_generation(&pool_address)
-            };
-
-            // Publish JetStream PoolCacheUpdate (authoritative SSOT).
-            // Reply Ok ONLY when JetStream write succeeds (I-24a).
-            let jetstream_ok = if let Some(ref nats) = ctx.nats {
-                let mut pool_update = PoolCacheUpdate::new_pool_discovered(
-                    "market-data",
-                    BUILD_VERSION,
-                    run_id,
-                    pool_address_str.clone(),
-                    "pump_amm".to_string(),
-                    base_mint_str.clone(),
-                    quote_mint_str,
-                    base_reserve,
-                    quote_reserve,
-                    None,
-                    0,
-                );
-                let mut meta = std::collections::HashMap::new();
-                let accounts_str: Vec<String> = accounts.iter().map(|p| p.to_string()).collect();
-                meta.insert("pool_accounts".to_string(), accounts_str.join(","));
-                meta.insert(
-                    "pump_amm_sell_cashback_remaining".to_string(),
-                    sell_flag_merged.to_string(),
-                );
-                meta.insert(
-                    "pump_amm_sell_layout_ready".to_string(),
-                    sell_layout_ready.to_string(),
-                );
-                if force_refresh {
-                    meta.insert(
-                        "pump_amm_sell_layout_authoritative".to_string(),
-                        "true".to_string(),
-                    );
-                }
-                meta.insert(
-                    "pump_amm_layout_generation".to_string(),
-                    layout_generation.to_string(),
-                );
-                if requires_fee_tail_merged {
-                    meta.insert(
-                        "pump_amm_sell_requires_fee_tail".to_string(),
-                        "true".to_string(),
-                    );
-                }
-                if requires_pre_fee_metas_merged {
-                    meta.insert(
-                        "pump_amm_sell_requires_pre_fee_metas".to_string(),
-                        "true".to_string(),
-                    );
-                }
-                if let Some(pk) = pre_fee_meta_1_merged {
-                    meta.insert("pump_amm_sell_pre_fee_meta_1".to_string(), pk.to_string());
-                }
-                if let Some(pk) = third_merged {
-                    meta.insert(
-                        "pump_amm_sell_cashback_third_meta".to_string(),
-                        pk.to_string(),
-                    );
-                }
-                if let Some(pk) = tail0_merged {
-                    meta.insert("pump_amm_sell_extended_tail_0".to_string(), pk.to_string());
-                }
-                if let Some(pk) = tail1_merged {
-                    meta.insert("pump_amm_sell_extended_tail_1".to_string(), pk.to_string());
-                }
-                if let Some(pk) = fee_tail0_merged {
-                    meta.insert(
-                        "pump_amm_sell_extended_fee_tail_0".to_string(),
-                        pk.to_string(),
-                    );
-                }
-                if let Some(pk) = fee_tail1_merged {
-                    meta.insert(
-                        "pump_amm_sell_extended_fee_tail_1".to_string(),
-                        pk.to_string(),
-                    );
-                }
-                if let Some(creator) = creator_opt {
-                    meta.insert("creator".to_string(), creator.to_string());
-                }
-                pool_update.metadata = Some(meta);
-                pool_update.set_dex_readiness_in_metadata(dex_readiness);
-                let subject = pool_subject(&pool_address_str);
-                match nats.jetstream_publish(&subject, &pool_update).await {
-                    Ok(true) => {
-                        let sell_ix_account_count = pump_amm_inferred_sell_ix_account_count(
-                            requires_pre_fee_metas_merged,
-                            requires_fee_tail_merged,
-                            sell_flag_merged,
-                        );
-                        info!(
-                            pool = %pool_address_str,
-                            base_mint = %base_mint_str,
-                            force_refresh,
-                            sell_requires_pre_fee_metas = requires_pre_fee_metas_merged,
-                            sell_pre_fee_meta_1 = ?pre_fee_meta_1_merged,
-                            sell_ix_account_count,
-                            sell_layout_ready,
-                            "EnsurePumpAmmPoolAccounts: Published PoolCacheUpdate to JetStream"
-                        );
-                        true
-                    }
-                    Ok(false) => {
-                        warn!(
-                            "EnsurePumpAmmPoolAccounts: JetStream publish failed (timeout or drop)"
-                        );
-                        false
-                    }
-                    Err(e) => {
-                        warn!(error = %e, "EnsurePumpAmmPoolAccounts: Failed to publish PoolCacheUpdate to JetStream");
-                        false
-                    }
-                }
-            } else {
-                false
-            };
-
-            if let Some(ref nats) = ctx.nats {
-                let (status, message) = pump_amm_control_response_for_ensure_publish(
-                    force_refresh,
-                    jetstream_ok,
-                    sell_layout_ready,
-                );
-                if force_refresh && jetstream_ok && !sell_layout_ready {
-                    log_pump_amm_scope49_force_refresh_sell_layout_failure(
-                        request_id,
-                        &pool_address_str,
-                        &base_mint_str,
-                        pump_amm_force_refresh_sell_diag.as_ref(),
-                    );
-                }
-                match (jetstream_ok, force_refresh, sell_layout_ready) {
-                    (true, _, true) => {
-                        info!(
-                            request_id = %request_id,
-                            base_mint = %base_mint_str,
-                            pool_address = %pool_address_str,
-                            rpc_global_scan = "no",
-                            control_response = "pending_publish",
-                            "I-24d Discovery: terminal outcome ok"
-                        );
-                    }
-                    (true, true, false) => {
-                        warn!(
-                            request_id = %request_id,
-                            base_mint = %base_mint_str,
-                            pool_address = %pool_address_str,
-                            sell_cashback_remaining = sell_flag_merged,
-                            sell_cashback_third_meta = ?third_merged,
-                            sell_extended_tail_0 = ?tail0_merged,
-                            sell_extended_tail_1 = ?tail1_merged,
-                            "I-24d Discovery: force_refresh result published as Partial (authoritative SELL layout unresolved)"
-                        );
-                    }
-                    (true, false, false) => {
-                        info!(
-                            request_id = %request_id,
-                            base_mint = %base_mint_str,
-                            pool_address = %pool_address_str,
-                            sell_cashback_remaining = sell_flag_merged,
-                            sell_cashback_third_meta = ?third_merged,
-                            sell_extended_tail_0 = ?tail0_merged,
-                            sell_extended_tail_1 = ?tail1_merged,
-                            "I-24d Discovery: terminal outcome ok (non-force-refresh publish may remain Partial)"
-                        );
-                    }
-                    (false, _, _) => {
-                        warn!(
-                            request_id = %request_id,
-                            base_mint = %base_mint_str,
-                            "I-24d Discovery: terminal outcome error (JetStream publish failed)"
-                        );
-                    }
-                }
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    status,
-                    Some(pool_address_str),
-                    message,
-                )
-                .await;
-            }
-        }
-        Ok(Some(_)) => {
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                outcome = "error",
-                reason = "pool_accounts_incomplete",
-                "I-24d Discovery: terminal outcome error (pool_accounts incomplete <14)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some("pool_accounts incomplete".to_string()),
-                )
-                .await;
-            }
-        }
-        Ok(None) => {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                outcome = "not_found",
-                "I-24d Discovery: terminal outcome not_found"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::NotFound,
-                    None,
-                    None,
-                )
-                .await;
-            }
-        }
-        Err(e) => {
-            warn!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                outcome = "error",
-                error = %e,
-                "I-24d Discovery: terminal outcome error (discovery failed)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-        }
-    }
-}
-
-/// Cold-path recovery: fetch PumpFun bonding curve account via RPC (force_refresh), update MASTER
-/// cache, publish JetStream PoolCacheUpdate. Does not short-circuit on cache when `force_refresh`.
-async fn handle_ensure_pumpfun_bonding_curve(
-    ctx: &MarketDataContext,
-    rpc: &Arc<SolanaRpc>,
-    run_id: &str,
-    request_id: &str,
-    base_mint_str: &str,
-    pool_address_hint: Option<&str>,
-    force_refresh: bool,
-) {
-    if defer_discovery_if_md_state_pressure(ctx, request_id).await {
-        return;
-    }
-    let base_mint = match Pubkey::from_str(base_mint_str) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!(base_mint = %base_mint_str, error = %e, "EnsurePumpfunBondingCurve: invalid base_mint");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let pumpfun = match PumpFunDex::new(Arc::clone(rpc), Some(ctx.live_pool_cache.clone())) {
-        Ok(d) => d,
-        Err(e) => {
-            warn!(error = %e, "EnsurePumpfunBondingCurve: PumpFunDex init failed");
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    None,
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let (bonding_curve, _) = pumpfun.derive_bonding_curve(&base_mint);
-    let bonding_curve = pool_address_hint
-        .and_then(|s| Pubkey::from_str(s).ok())
-        .unwrap_or(bonding_curve);
-    let bonding_curve_str = bonding_curve.to_string();
-
-    if !force_refresh {
-        if let Some(CachedPoolState::PumpFun(_)) = ctx.live_pool_cache.get(&bonding_curve) {
-            info!(
-                request_id = %request_id,
-                base_mint = %base_mint_str,
-                pool = %bonding_curve_str,
-                "EnsurePumpfunBondingCurve: cache hit (skip RPC)"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Ok,
-                    Some(bonding_curve_str),
-                    None,
-                )
-                .await;
-            }
-            return;
-        }
-    } else {
-        warn!(
-            request_id = %request_id,
-            base_mint = %base_mint_str,
-            pool = %bonding_curve_str,
-            "EnsurePumpfunBondingCurve: force_refresh — always fetching bonding curve via RPC (no cache-first)"
-        );
-    }
-
-    let state = match rpc.get_account_retry(&bonding_curve).await {
-        Ok(acct) => match BondingCurveState::parse(&acct.data) {
-            Ok(s) => s,
-            Err(e) => {
-                warn!(
-                    request_id = %request_id,
-                    pool = %bonding_curve_str,
-                    error = %e,
-                    "EnsurePumpfunBondingCurve: parse bonding curve failed"
-                );
-                if let Some(ref nats) = ctx.nats {
-                    publish_control_response(
-                        nats,
-                        &ctx.run_id,
-                        request_id,
-                        ControlResponseStatus::Error,
-                        Some(bonding_curve_str),
-                        Some(format!("parse error: {e}")),
-                    )
-                    .await;
-                }
-                return;
-            }
-        },
-        Err(e) => {
-            warn!(
-                request_id = %request_id,
-                pool = %bonding_curve_str,
-                error = %e,
-                "EnsurePumpfunBondingCurve: RPC get_account failed"
-            );
-            if let Some(ref nats) = ctx.nats {
-                publish_control_response(
-                    nats,
-                    &ctx.run_id,
-                    request_id,
-                    ControlResponseStatus::Error,
-                    Some(bonding_curve_str),
-                    Some(e.to_string()),
-                )
-                .await;
-            }
-            return;
-        }
-    };
-
-    let token_program = ctx
-        .live_pool_cache
-        .get_mint_program(&base_mint)
-        .unwrap_or_else(|| Pubkey::new_from_array(spl_token::id().to_bytes()));
-    let (associated_bonding_curve, _) =
-        pumpfun.derive_associated_bonding_curve(&bonding_curve, &base_mint, &token_program);
-
-    let cached = CachedPoolState::PumpFun(PumpFunState {
-        token_mint: base_mint,
-        bonding_curve,
-        associated_bonding_curve,
-        virtual_sol_reserves: state.virtual_sol_reserves,
-        virtual_token_reserves: state.virtual_token_reserves,
-        real_sol_reserves: state.real_sol_reserves,
-        real_token_reserves: state.real_token_reserves,
-        complete: state.complete,
-        creator: state.creator,
-        cashback_enabled: state.cashback_enabled,
-    });
-    ctx.live_pool_cache.upsert(bonding_curve, cached, 0);
-
-    let jetstream_ok = if let Some(ref nats) = ctx.nats {
-        let mut pool_update = PoolCacheUpdate::new_pool_discovered(
-            "market-data",
-            BUILD_VERSION,
-            run_id,
-            bonding_curve_str.clone(),
-            "pumpfun".to_string(),
-            base_mint.to_string(),
-            NATIVE_SOL_MINT.to_string(),
-            state.virtual_token_reserves,
-            state.virtual_sol_reserves,
-            Some(0),
-            0,
-        );
-        let mut meta = std::collections::HashMap::new();
-        meta.insert("creator".to_string(), state.creator.to_string());
-        meta.insert(
-            "associated_bonding_curve".to_string(),
-            associated_bonding_curve.to_string(),
-        );
-        meta.insert("complete".to_string(), state.complete.to_string());
-        meta.insert(
-            "real_token_reserves".to_string(),
-            state.real_token_reserves.to_string(),
-        );
-        meta.insert(
-            "real_sol_reserves".to_string(),
-            state.real_sol_reserves.to_string(),
-        );
-        meta.insert(
-            "cashback_enabled".to_string(),
-            state.cashback_enabled.to_string(),
-        );
-        pool_update.metadata = Some(meta);
-        // Cold-path RPC refresh: explicit Ready for JetStream / SLAVE (Bug #36).
-        pool_update.set_dex_readiness_in_metadata(DexPoolReadiness::Ready);
-        ctx.live_pool_cache
-            .merge_pumpfun_bonding_readiness(bonding_curve, DexPoolReadiness::Ready);
-        let subject = pool_subject(&bonding_curve_str);
-        match nats.jetstream_publish(&subject, &pool_update).await {
-            Ok(true) => {
-                info!(
-                    pool = %bonding_curve_str,
-                    base_mint = %base_mint_str,
-                    "EnsurePumpfunBondingCurve: Published PoolCacheUpdate to JetStream"
-                );
-                true
-            }
-            Ok(false) => {
-                warn!("EnsurePumpfunBondingCurve: JetStream publish failed (timeout or drop)");
-                false
-            }
-            Err(e) => {
-                warn!(
-                    error = %e,
-                    "EnsurePumpfunBondingCurve: Failed to publish PoolCacheUpdate to JetStream"
-                );
-                false
-            }
-        }
-    } else {
-        false
-    };
-
-    if let Some(ref nats) = ctx.nats {
-        let (status, message) = if jetstream_ok {
-            (ControlResponseStatus::Ok, None)
-        } else {
-            (
-                ControlResponseStatus::Error,
-                Some("JetStream publish failed".to_string()),
-            )
-        };
-        publish_control_response(
-            nats,
-            &ctx.run_id,
-            request_id,
-            status,
-            Some(bonding_curve_str),
-            message,
-        )
-        .await;
-    }
-}
-
-async fn publish_control_response(
-    nats: &NatsClient,
-    run_id: &str,
-    request_id: &str,
-    status: ControlResponseStatus,
-    pool_address: Option<String>,
-    message: Option<String>,
-) {
-    let mut resp = ControlResponse::new(
-        "market-data",
-        BUILD_VERSION,
-        run_id,
-        request_id.to_string(),
-        "market-data",
-        status,
-    );
-    let pool_for_log = pool_address.clone();
-    if let Some(pa) = pool_address {
-        resp = resp.with_pool_address(pa);
-    }
-    if let Some(m) = message {
-        resp = resp.with_message(m);
-    }
-    let status_str = format!("{:?}", status);
-    if let Err(e) = nats.publish(TOPIC_CONTROL_RESPONSES, &resp).await {
-        warn!(
-            request_id = %request_id,
-            status = %status_str,
-            pool_address = ?pool_for_log,
-            error = %e,
-            "I-24d Discovery: Failed to publish ControlResponse"
-        );
-    } else {
-        info!(
-            request_id = %request_id,
-            status = %status_str,
-            pool_address = ?pool_for_log,
-            "I-24d Discovery: ControlResponse published"
-        );
-    }
 }
 
 /// Phase 5c: md-account-publish runtime (delegates to `publish/account.rs`).
@@ -11991,7 +8392,7 @@ async fn run_geyser_loop(
                                         let dex_clone = Arc::clone(&pump_amm_dex);
                                         tokio::spawn(async move {
                                             handle_ensure_pump_amm_pool_accounts(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 dex_clone.as_ref(),
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12018,7 +8419,7 @@ async fn run_geyser_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_pumpfun_bonding_curve(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12045,7 +8446,7 @@ async fn run_geyser_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_orca_whirlpool_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12072,7 +8473,7 @@ async fn run_geyser_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_meteora_dlmm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12099,7 +8500,7 @@ async fn run_geyser_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_raydium_amm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12126,7 +8527,7 @@ async fn run_geyser_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_raydium_cpmm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12153,7 +8554,7 @@ async fn run_geyser_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_meteora_cpmm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12837,7 +9238,7 @@ async fn run_simulation_loop(
                                         let dex_clone = Arc::clone(&pump_amm_dex);
                                         tokio::spawn(async move {
                                             handle_ensure_pump_amm_pool_accounts(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 dex_clone.as_ref(),
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12864,7 +9265,7 @@ async fn run_simulation_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_pumpfun_bonding_curve(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12891,7 +9292,7 @@ async fn run_simulation_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_orca_whirlpool_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12918,7 +9319,7 @@ async fn run_simulation_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_meteora_dlmm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12945,7 +9346,7 @@ async fn run_simulation_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_raydium_amm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12972,7 +9373,7 @@ async fn run_simulation_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_raydium_cpmm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -12999,7 +9400,7 @@ async fn run_simulation_loop(
                                         let rpc_clone = rpc.clone();
                                         tokio::spawn(async move {
                                             handle_ensure_meteora_cpmm_pool_state(
-                                                &ctx_clone,
+                                                ctx_clone.as_ref(),
                                                 &rpc_clone,
                                                 run_id.as_str(),
                                                 &request_id,
@@ -13116,7 +9517,15 @@ async fn run_simulation_loop(
 #[cfg(test)]
 mod discovery_tests {
     use super::*;
-    use ironcrab::execution::live_pool_cache::RaydiumAmmState;
+    use ironcrab::execution::live_pool_cache::{
+        meteora_dlmm_readiness_for_pool_cache_update, orca_readiness_for_pool_cache_update,
+        raydium_amm_readiness_for_pool_cache_update, RaydiumAmmState,
+    };
+    use ironcrab::ipc::ControlResponseStatus;
+    use ironcrab::market_data::cold::{
+        pump_amm_control_response_for_ensure_publish, pump_amm_sell_layout_publish_state,
+        pump_amm_sell_layout_state_for_ensure_publish,
+    };
 
     /// Positive Ack ONLY when JetStream write succeeds (I-24a).
     /// This helper encodes the invariant; the handler uses the same logic.
@@ -17854,6 +14263,43 @@ mod pr_b_geyser_tracking_tests {
         assert!(
             bin_src.contains("fn spawn_market_data_account_publish_runtime"),
             "bin must retain thin spawn wrapper for md-publish runtime"
+        );
+    }
+
+    /// Phase 5d: I-24d Ensure* cold-path handlers live in dedicated library module.
+    #[test]
+    fn phase5d_cold_ensure_in_dedicated_module() {
+        let ensure_pump_src = include_str!("../market_data/cold/ensure_pump.rs");
+        let rpc_src = include_str!("../market_data/cold/rpc_refresh.rs");
+        let host_src = include_str!("../market_data/cold/host.rs");
+        let bin_src = include_str!("market_data.rs");
+        let prod_src = bin_src
+            .split("#[cfg(test)]")
+            .next()
+            .expect("production source section");
+        assert!(
+            ensure_pump_src.contains("pub async fn handle_ensure_pump_amm_pool_accounts"),
+            "ensure_pump.rs must contain handle_ensure_pump_amm_pool_accounts"
+        );
+        assert!(
+            rpc_src.contains("pub async fn cold_path_rpc_refresh_orca_whirlpool_pool_row"),
+            "rpc_refresh.rs must contain cold_path_rpc_refresh_orca_whirlpool_pool_row"
+        );
+        assert!(
+            host_src.contains("pub trait ColdHost"),
+            "host.rs must define ColdHost"
+        );
+        assert!(
+            bin_src.contains("use ironcrab::market_data::cold::"),
+            "market_data bin must import cold module"
+        );
+        assert!(
+            bin_src.contains("impl ColdHost for MarketDataContext"),
+            "bin must wire MarketDataContext to ColdHost"
+        );
+        assert!(
+            !prod_src.contains("async fn handle_ensure_pump_amm_pool_accounts("),
+            "Ensure handlers must not remain defined in bin production code"
         );
     }
