+++ src/market_data/mod.rs
@@ -1,3 +1,4 @@
+pub mod cold;
 pub mod publish;
 pub mod sidefx;
 pub mod track;
